### 2020-10-28 #2794 @whedon
[PRE REVIEW]: effectsize: Estimation and Interpretation of Effect Size Indices and Standardized Parameters
**Submitting author:** @mattansb (<a href="http://orcid.org/0000-0002-4287-4801">Mattan S. Ben-Shachar</a>)
**Repository:** <a href="https://github.com/easystats/effectsize" target ="_blank">https://github.com/easystats/effectsize</a>
**Version:** 0.4.0.001
