### 2020-10-28 #2794 @mattansb
@arfon An alternative title would be "effectsize: Estimation and Interpretation of Effect Size Indices and Standardized Parameters".
Would that be acceptable?

For reviewers, we would like to suggest Guy Prochilo (gprochilo), Tristan Mahr (tjmahr) or Matt Craddock (craddm).

Thanks!
