### 2020-11-03 b01bbc4
only accept linear model for response standardization
