+++ R/standardize.models.R
@@ -226,5 +226,6 @@ standardize.wbgee <- standardize.wbm
 #' @keywords internal
 .no_response_standardize <- function(info) {
   # check if model has a response variable that should not be standardized.
-  info$is_count | info$is_ordinal | info$is_multinomial | info$is_beta | info$is_censored | info$is_binomial | info$is_survival
+  info$is_linear & info$family != "inverse.gaussian"
+  # info$is_count | info$is_ordinal | info$is_multinomial | info$is_beta | info$is_censored | info$is_binomial | info$is_survival
 }
