+++ R/normalize.R
@@ -23,7 +23,7 @@
 #' @references
 #' - Smithson M, Verkuilen J (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, 11(1), 54–71.
 #'
-#' @seealso [ranktransform()] [standardize()] [change_scale()]
+#' @family transform utilities
 #'
 #' @return A normalized object.
 #' @export
