+++ man/ranktransform.Rd
@@ -59,5 +59,9 @@ head(ranktransform(iris))
 
 }
 \seealso{
-\code{\link[=normalize]{normalize()}} \code{\link[=standardize]{standardize()}} \code{\link[=change_scale]{change_scale()}}
+Other transform utilities: 
+\code{\link{change_scale}()},
+\code{\link{normalize}()},
+\code{\link{standardize}()}
 }
+\concept{transform utilities}
