### 2020-11-03 bb2dd78
family convert between effect sizes
