+++ man/oddsratio_to_riskratio.Rd
@@ -33,3 +33,10 @@ oddsratio_to_riskratio(OR, p0 = p0)
 \references{
 Grant, R. L. (2014). Converting an odds ratio to a range of plausible relative risks for better communication of research findings. Bmj, 348, f7450.
 }
+\seealso{
+Other convert between effect sizes: 
+\code{\link{d_to_common_language}()},
+\code{\link{d_to_r}()},
+\code{\link{odds_to_probs}()}
+}
+\concept{convert between effect sizes}
