+++ R/standardize_parameters.R
@@ -83,6 +83,7 @@
 #' @return A data frame with the standardized parameters and their CIs.
 #'
 #' @family standardize
+#' @family effect size indices
 #'
 #' @examples
 #' library(effectsize)
