+++ man/standardize_parameters.Rd
@@ -191,5 +191,13 @@ if (require("rstanarm")) {
 Other standardize: 
 \code{\link{standardize_info}()},
 \code{\link{standardize}()}
+
+Other effect size indices: 
+\code{\link{cohens_d}()},
+\code{\link{effectsize}()},
+\code{\link{eta_squared_posterior}()},
+\code{\link{eta_squared}()},
+\code{\link{phi}()}
 }
+\concept{effect size indices}
 \concept{standardize}
