+++ man/eta_squared_posterior.Rd
@@ -70,3 +70,12 @@ if (require(car)) {
 
 
 }
+\seealso{
+Other effect size indices: 
+\code{\link{cohens_d}()},
+\code{\link{effectsize}()},
+\code{\link{eta_squared}()},
+\code{\link{phi}()},
+\code{\link{standardize_parameters}()}
+}
+\concept{effect size indices}
