### 2020-11-03 bc59781
family effect size from test statistic
