### 2020-11-03 be9b210
Added CI section (inherited)
