+++ man/eta_squared.Rd
@@ -82,18 +82,6 @@ and \emph{centered covariates}, for sensible results. See examples and the \code
 package.
 }
 
-\subsection{Confidence Intervals}{
-
-Confidence intervals are estimated using the Noncentrality parameter method;
-These methods searches for a the best \code{ncp} (non-central parameters) for
-of the noncentral F distribution for the desired tail-probabilities,
-and then convert these \code{ncp}s to the corresponding effect sizes.
-\cr\cr
-Special care should be taken when interpreting CIs with a lower bound equal
-to (or small then) 0, and even more care should be taken when the \emph{upper}
-bound is equal to (or small then) 0 (Steiger, 2004; Morey et al., 2016).
-}
-
 \subsection{Un-Biased Estimate of Eta}{
 
 Both \emph{\strong{Omega}} and \emph{\strong{Epsilon}} are unbiased estimators of the
@@ -119,6 +107,21 @@ Cohen has suggested that the values of 0.10, 0.25, and 0.40 represent small,
 medium, and large effect sizes, respectively.
 }
 }
+\section{Confidence Intervals}{
+Confidence intervals are estimated using the Noncentrality parameter method;
+These methods searches for a the best non-central parameters (\code{ncp}s) of the
+noncentral t-, F- or Chi-squared distribution for the desired
+tail-probabilities, and then convert these \code{ncp}s to the corresponding effect
+sizes.
+}
+
+\section{CI Contains Zero}{
+
+Special care should be taken when interpreting CIs with a lower bound equal
+to (or small then) 0, and even more care should be taken when the \emph{upper}
+bound is equal to (or small then) 0 (Steiger, 2004; Morey et al., 2016).
+}
+
 \examples{
 \donttest{
 library(effectsize)
