### 2020-11-04 #2815 @whedon
[REVIEW]: effectsize: Estimation of Effect Size Indices and Standardized Parameters
**Submitting author:** @mattansb (<a href="http://orcid.org/0000-0002-4287-4801">Mattan S. Ben-Shachar</a>)
**Repository:** <a href="https://github.com/easystats/effectsize" target ="_blank">https://github.com/easystats/effectsize</a>
**Version:** v0.4.1.2
