### 2020-11-04 #2815 @arfon
> Can't find any papers to compile :-(

Lemme dig into this @mikldk - not sure what is going on here.
