### 2020-11-04 #2815 @mikldk
@mattansb (and possibly @arfon): When I look into the [repo](https://github.com/easystats/effectsize), it is not clear if the software submitted is at the master branch or in the paper branch?
