### 2020-11-04 #2815 @arfon
> @mattansb (and possibly @arfon): When I look into the repo, it is not clear if the software submitted is at the master branch or in the paper branch?

I think we're good now ☝️. I accidentally introduced a regression into Whedon last night.
