### 2020-11-04 #2815 @arfon
> @mattansb: In the paper branch there is a copy of the R package itself and not just the paper for JOSS. @arfon, shouldn't this copy of the R package in the paper branch be removed such that only the paper etc. are in the paper branch?

We generally don't insist on this but it would be good to clarify for the reviewers here which branch of the _software_ should be reviewed.
