### 2020-11-04 #2815 @mattansb
@arfon & @mikldk The paper branch has a copy of the current CRAN version of the package. But I think it will reduce ambiguity if I just merge the paper branch into master, so I'll do that now. Sorry about the confusion.
