### 2020-11-04 #2815 @mikldk
> @mikldk Reviewers should look at the `master` branch which now has both the code _and_ the paper.

Thanks, @mattansb . You should consider deleting the `paper` branch.
