### 2020-11-04 #2815 @mikldk
@trashbirdecology, @jkarreth: Once again, thanks for agreeing to review. The problems are now resolved, and the `master` branch is ready for your reviews. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

(@mattansb, @jkarreth: Note that we have a slightly extended time frame at the moment, and @jkarreth do not expect to be able to complete their review before mid-December, but will try if at all possible.)
