### 2020-11-04 ab70231
Add right argument to rules
Making inclusion threshold more explicit
#194
