### 2020-11-04 31bbdff
add threshold-right to vignette
#194
