+++ vignettes/interpret.Rmd
@@ -54,11 +54,11 @@ interpret_r(x, rules = "funder2019")
 ```
 
 - **r < 0.05** - Tiny
-- **0.05 < r < 0.1** - Very small
-- **0.1 < r < 0.2** - Small
-- **0.2 < r < 0.3** - Medium
-- **0.3 < r < 0.4** - Large
-- **r > 0.4** - Very large
+- **0.05 <= r < 0.1** - Very small
+- **0.1 <= r < 0.2** - Small
+- **0.2 <= r < 0.3** - Medium
+- **0.3 <= r < 0.4** - Large
+- **r >= 0.4** - Very large
 
 
 
@@ -71,9 +71,9 @@ interpret_r(x, rules = "gignac2016")
 ```
 
 - **r < 0.1** - Very small
-- **0.1 < r < 0.2** - Small
-- **0.2 < r < 0.3** - Moderate
-- **r > 0.3** - Large
+- **0.1 <= r < 0.2** - Small
+- **0.2 <= r < 0.3** - Moderate
+- **r >= 0.3** - Large
 
 
 
@@ -84,9 +84,9 @@ interpret_r(x, rules = "cohen1988")
 ```
 
 - **r < 0.1** - Very small
-- **0.1 < r < 0.3** - Small
-- **0.3 < r < 0.5** - Moderate
-- **r > 0.5** - Large
+- **0.1 <= r < 0.3** - Small
+- **0.3 <= r < 0.5** - Moderate
+- **r >= 0.5** - Large
 
 
 #### @evans1996straightforward
@@ -96,10 +96,10 @@ interpret_r(x, rules = "evans1996")
 ```
 
 - **r < 0.2** - Very weak
-- **0.2 < r < 0.4** - Weak
-- **0.4 < r < 0.6** - Moderate
-- **0.6 < r < 0.8** - Strong
-- **r > 0.8** - Very strong
+- **0.2 <= r < 0.4** - Weak
+- **0.4 <= r < 0.6** - Moderate
+- **0.6 <= r < 0.8** - Strong
+- **r >= 0.8** - Very strong
 
 
 
@@ -114,9 +114,9 @@ interpret_d(x, rules = "cohen1988")
 ```
 
 - **d < 0.2** - Very small
-- **0.2 < d < 0.5** - Small
-- **0.5 < d < 0.8** - Medium
-- **d > 0.8** - Large
+- **0.2 <= d < 0.5** - Small
+- **0.5 <= d < 0.8** - Medium
+- **d >= 0.8** - Large
 
 
 #### @sawilowsky2009new
@@ -126,12 +126,12 @@ interpret_d(x, rules = "sawilowsky2009")
 ```
 
 - **d < 0.1** - Tiny
-- **0.1 < d < 0.2** - Very small
-- **0.2 < d < 0.5** - Small
-- **0.5 < d < 0.8** - Medium
-- **0.8 < d < 1.2** - Large
-- **1.2 < d < 2** - Very large
-- **d > 2** - Huge
+- **0.1 <= d < 0.2** - Very small
+- **0.2 <= d < 0.5** - Small
+- **0.5 <= d < 0.8** - Medium
+- **0.8 <= d < 1.2** - Large
+- **1.2 <= d < 2** - Very large
+- **d >= 2** - Huge
 
 
 #### @gignac2016effect
@@ -143,9 +143,9 @@ interpret_d(x, rules = "gignac2016")
 ```
 
 - **d < 0.2** - Very small
-- **0.2 < d < 0.41** - Small
-- **0.41 < d < 0.63** - Moderate
-- **d > 0.63** - Large
+- **0.2 <= d < 0.41** - Small
+- **0.41 <= d < 0.63** - Moderate
[…]
