### 2020-11-04 7c6e33c
clean up printing for reference values rules
