+++ DESCRIPTION
@@ -42,8 +42,8 @@ Imports:
 Suggests:
 	afex,
 	BayesFactor,
-	brms,
 	boot,
+	brms,
 	car,
 	correlation,
 	covr,
@@ -52,11 +52,12 @@ Suggests:
 	gamm4,
 	ggplot2,
 	knitr,
-	lmerTest,
 	lm.beta,
 	lme4,
-	MuMIn,
+	lmerTest,
+	mediation,
 	modelbased, 
+	MuMIn,
 	performance,
 	pscl,
 	rlang,
