+++ NAMESPACE
@@ -49,6 +49,7 @@ S3method(standardize,factor)
 S3method(standardize,grouped_df)
 S3method(standardize,integer)
 S3method(standardize,logical)
+S3method(standardize,mediate)
 S3method(standardize,mlm)
 S3method(standardize,numeric)
 S3method(standardize,wbgee)
