### 2020-11-05 #2815 @trashbirdecology
[Review issue opened at repository](https://github.com/easystats/effectsize/issues/196).
