### 2020-11-06 767a26f
add MANOVA for eta_sqaured
https://github.com/easystats/parameters/pull/326
