### 2020-11-07 0c9e8c1
added oddsratio and riskratio functions
