+++ man/phi.Rd
@@ -1,9 +1,11 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/phi_cramers_v.R
+% Please edit documentation in R/xtab.R
 \name{phi}
 \alias{phi}
 \alias{cohens_w}
 \alias{cramers_v}
+\alias{oddsratio}
+\alias{riskratio}
 \alias{cohens_g}
 \title{Effect size for contingency tables}
 \usage{
@@ -12,6 +14,10 @@ phi(x, y = NULL, ci = 0.95, adjust = FALSE, CI, ...)
 cohens_w(x, y = NULL, ci = 0.95, adjust = FALSE, CI, ...)
 
 cramers_v(x, y = NULL, ci = 0.95, adjust = FALSE, CI, ...)
+
+oddsratio(x, y = NULL, ci = 0.95, log = FALSE, ...)
+
+riskratio(x, y = NULL, ci = 0.95, log = FALSE, ...)
 
 cohens_g(x, y = NULL, ci = 0.95, ...)
 }
@@ -30,30 +36,42 @@ cohens_g(x, y = NULL, ci = 0.95, ...)
 
 \item{...}{Arguments passed to \code{\link[stats:chisq.test]{stats::chisq.test()}}, such as \code{p}. Ignored
 for \code{cohens_g()}.}
+
+\item{log}{Take in or output the log of the ratio (such as in logistic models).}
 }
 \value{
 A data frame with the effect size(s), and confidence interval(s).
 }
 \description{
-Compute Cramer's \emph{V}, phi (\eqn{\phi}), Cohen's \emph{w} (an alias of phi) and
-Cohen's \emph{g} for contingency tables or goodness-of-fit. See details.
+Compute Cramer's \emph{V}, phi (\eqn{\phi}), Cohen's \emph{w} (an alias of phi), Odds
+ratios, Risk ratios and Cohen's \emph{g} for contingency tables or
+goodness-of-fit. See details.
 }
 \details{
 Cramer's \emph{V} and phi (\eqn{\phi}) are effect sizes for tests of independence
-in 2D contingency tables, or for goodness-of-fit in 1D tables. For 2x2
-tables, all 3 are identical, and are equal to the simple correlation between
+in 2D contingency tables, or for goodness-of-fit in 1D tables. For 2-by-2
+tables, they are identical, and are equal to the simple correlation between
 two dichotomous variables, ranging between  0 (no dependence) and 1 (perfect
 dependence). For larger tables, Cramer's \emph{V} should be used, as it is bounded
 between 0-1, whereas phi can be larger than 1.
+\cr\cr
+For 2-by-2 contingency tables, Odds ratios and Risk ratios can also be
+estimated. Note that these are computed with each \strong{column} representing the
+different groups, and the first column representing the baseline (or
+control). If you wish you use rows as groups you must pass a transposed
+table, or switch the \code{x} and \code{y} arguments.
 \cr\cr
 Cohen's \emph{g} is an effect size for dependent (paired) contingency tables
 ranging between 0 (perfect symmetry) and 0.5 (perfect asymmetry) (see
 \code{\link[stats:mcnemar.test]{stats::mcnemar.test()}}).
 }
-\section{Confidence Intervals for g}{
+\section{Confidence Intervals for g, OR and RR}{
 For Cohen's \emph{g}, confidence intervals are based on the proportion (\eqn{P = g
 + 0.5}) confidence intervals returned by \code{\link[stats:prop.test]{stats::prop.test()}} (minus 0.5),
 which give a good close approximation.
+\cr\cr
[…]
