### 2020-11-08 cfa7639
improvments and changes to eta_squared_posterior
