+++ tests/testthat/test-eta_squared_posterior.R
@@ -21,7 +21,10 @@ if (require("testthat") &&
       partial = TRUE
     )
 
-    testthat::expect_warning(es_post <- eta_squared_posterior(fit_bayes, verbose = FALSE))
+    testthat::expect_warning(
+      es_post <- eta_squared_posterior(fit_bayes, verbose = FALSE,
+                                       ss_function = car::Anova, type = 3)
+    )
     testthat::expect_equal(colnames(es_post), es_tab$Parameter)
 
     # this is a very soft test...
@@ -30,24 +33,27 @@ if (require("testthat") &&
   })
 
 
-  test_that("eta_squared_posterior (non-PARTIAL, type = 2)", {
+  test_that("eta_squared_posterior (non-PARTIAL, type = 3)", {
     testthat::skip_on_cran()
 
     es_tab <- eta_squared(
       Anova(
         lm(mpg ~ factor(cyl) * wt + qsec,
            data = mtcars),
-        type = 2
+        type = 3
       ),
       partial = FALSE
     )
 
-    testthat::expect_warning(es_post <- eta_squared_posterior(fit_bayes, verbose = FALSE, partial = FALSE, type = 2))
+    testthat::expect_warning(
+      es_post <- eta_squared_posterior(fit_bayes, verbose = FALSE, partial = FALSE,
+                                       ss_function = car::Anova, type = 3)
+    )
     testthat::expect_equal(colnames(es_post), es_tab$Parameter)
 
     # this is a very soft test...
     es_tab_bayes <- describe_posterior(es_post)
-    # testthat::expect_equal(order(es_tab_bayes$Median), order(es_tab$Eta2))
+    testthat::expect_equal(order(es_tab_bayes$Median), order(es_tab$Eta2))
   })
 }
 
