+++ man/phi.Rd
@@ -139,7 +139,6 @@ cohens_g(Performance)
 Other effect size indices: 
 \code{\link{cohens_d}()},
 \code{\link{effectsize}()},
-\code{\link{eta_squared_posterior}()},
 \code{\link{eta_squared}()},
 \code{\link{standardize_parameters}()}
 }
