+++ man/eta_squared.Rd
@@ -1,11 +1,12 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/eta_squared.R
+% Please edit documentation in R/eta_squared.R, R/eta_squared_posterior.R
 \name{eta_squared}
 \alias{eta_squared}
 \alias{omega_squared}
 \alias{epsilon_squared}
 \alias{cohens_f}
 \alias{cohens_f_squared}
+\alias{eta_squared_posterior}
 \title{Effect size for ANOVA}
 \usage{
 eta_squared(model, partial = TRUE, generalized = FALSE, ci = 0.9, ...)
@@ -24,6 +25,16 @@ cohens_f_squared(
   model2 = NULL,
   ...
 )
+
+eta_squared_posterior(
+  model,
+  partial = TRUE,
+  generalized = FALSE,
+  ss_function = stats::anova,
+  draws = 500,
+  verbose = TRUE,
+  ...
+)
 }
 \arguments{
 \item{model}{A model, ANOVA object, or the result of \code{parameters::model_parameters}.}
@@ -39,15 +50,29 @@ automatically from the fitted model, if they were provided then.}
 
 \item{ci}{Confidence Interval (CI) level}
 
-\item{...}{Arguments passed to or from other methods (ignored).}
+\item{...}{For Bayesian models, passed to \code{ss_function}. Otherwise ignored.}
 
 \item{squared}{Return Cohen's \emph{f} or Cohen's \emph{f}-squared?}
 
-\item{model2}{Second model. If specified, returns Cohen's \emph{f} for
-R-squared-change between the two models.}
+\item{model2}{Optional second model for Cohen's f (/squared). If specified,
+returns the effect size for R-squared-change between the two models.}
+
+\item{ss_function}{For Bayesian models, the function used to extract
+sum-of-squares. uses \code{\link[=anova]{anova()}} by default, but can also be \code{\link[car:Anova]{car::Anova()}}
+for simple linear models.}
+
+\item{draws}{For Bayesian models, an integer indicating the number of draws
+from the posterior predictive distribution to return. Larger numbers take
+longer to run, but provide estimates that are more stable.}
+
+\item{verbose}{Show messages / warning about centering.}
 }
 \value{
 A data frame with the effect size(s) and confidence interval(s).
+\cr\cr
+For \code{eta_squared_posterior()}, a data frame containing the ppd of the Eta
+squared for each fixed effect, which can then be passed to
+\code{\link[bayestestR:describe_posterior]{bayestestR::describe_posterior()}} for summary stats.
 
 A data frame containing the effect size values and their confidence intervals.
 }
@@ -106,6 +131,17 @@ and the \\% unexplained variance.
 Cohen has suggested that the values of 0.10, 0.25, and 0.40 represent small,
 medium, and large effect sizes, respectively.
 }
+
+\subsection{Eta Squared from Posterior Predictive Distribution}{
+
+For Bayesian models (fit with \code{brms} or \code{rstanarm}),
+\code{eta_squared_posterior()} simulates data from the posterior predictive
+distribution (ppd) and for each simulation the Eta Squared is computed for
+the model's fixed effects. This means that the returned values are the
+population level effect size as implied by the posterior model (and not the
+effect size in the sample data). See \code{\link[rstantools:posterior_predict]{rstantools::posterior_predict()}} for
+more info.
+}
 }
 \section{Confidence Intervals}{
 Confidence intervals are estimated using the Noncentrality parameter method;
[…]

[…]
