+++ R/effectsize.R
@@ -26,6 +26,9 @@
 #' @family effect size indices
 #'
 #' @examples
+#'
+#' ## Hypothesis Testing
+#' ## ------------------
 #' contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))
 #' Xsq <- chisq.test(contingency_table)
 #' effectsize(Xsq)
@@ -36,6 +39,9 @@
 #' Aov <- oneway.test(extra ~ group, data = sleep)
 #' effectsize(Aov)
 #'
+#'
+#' ## Bayesian Hypothesis Testing
+#' ## ---------------------------
 #' if (require(BayesFactor)) {
 #'   bf1 <- ttestBF(mtcars$mpg[mtcars$am == 1], mtcars$mpg[mtcars$am == 0])
 #'   effectsize(bf1, test = NULL)
@@ -48,6 +54,9 @@
 #'   effectsize(bf3, test = NULL)
 #' }
 #'
+#'
+#' ## Models and Anova Tables
+#' ## -----------------------
 #' fit <- lm(mpg ~ factor(cyl) * wt + hp, data = mtcars)
 #' effectsize(fit)
 #'
