### 2020-11-09 49633ce
https://github.com/easystats/report/issues/109
