+++ R/standardize.models.R
@@ -293,7 +293,7 @@ standardize.wbgee <- standardize.wbm
   x <- insight::find_terms(model, flatten = TRUE)
   # log_pattern <- "^log\\((.*)\\)"
   log_pattern <- "(log\\(log|log|log1|log10|log1p|log2)\\(([^,\\+)]*).*"
-  out <- gsub(log_pattern, "\\2", x[grepl(log_pattern, x)])
+  out <- trimws(gsub(log_pattern, "\\2", x[grepl(log_pattern, x)]))
   intersect(colnames(data), out)
 }
 
