### 2020-11-10 bbec0d9
testthat 3rd ed
tol -> tolerance (and fix test values)
check.attributes -> ignore_attr
