+++ tests/testthat/test-standardize_parameters.R
@@ -8,7 +8,7 @@ if (require("testthat") && require("effectsize")) {
 
     model <- lm(Sepal.Length ~ Petal.Length, data = df)
     es <- standardize_parameters(model)
-    testthat::expect_equal(es[2,2], r, tol = 0.01)
+    testthat::expect_equal(es[2,2], r, tolerance = 0.01)
   })
 
 
@@ -43,58 +43,58 @@ if (require("testthat") && require("effectsize")) {
     testthat::expect_equal(
       standardize_parameters(model, method = "refit")$Std_Coefficient,
       c(0.044, -0.072, -0.060, 0.844),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, method = "posthoc")$Std_Coefficient,
       c(0, -0.072, -0.060, 0.844),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, method = "smart")$Std_Coefficient,
       c(0, -0.170, -0.142, 0.844),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, method = "basic")$Std_Coefficient,
       c(0, -0.034, -0.028, 0.844),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     ## CI
     testthat::expect_equal(
       standardize_parameters(model, method = "basic")$CI_low,
       c(0, -0.294, -0.433, 0.491),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, method = "basic")$CI_high,
       c(0, 0.225, 0.375, 1.196),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, ci = 0.8, method = "basic")$CI_low,
       c(0, -0.203, -0.292, 0.614),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     testthat::expect_equal(
       standardize_parameters(model, ci = 0.8, method = "basic")$CI_high,
       c(0, 0.135, 0.234, 1.073),
-      tol = 0.01
+      tolerance = 0.01
     )
 
     data("mtcars")
     m0 <- lm(mpg ~ cyl + factor(am), mtcars)
     testthat::expect_equal(standardize_parameters(m0, method = "refit")[[2]][-1],
-                           standardize_parameters(m0, method = "smart")[[2]][-1], tol = 0.01)
+                           standardize_parameters(m0, method = "smart")[[2]][-1], tolerance = 0.01)
     testthat::expect_equal(standardize_parameters(m0, method = "refit", two_sd = TRUE)[[2]][-1],
-                           standardize_parameters(m0, method = "smart", two_sd = TRUE)[[2]][-1], tol = 0.01)
+                           standardize_parameters(m0, method = "smart", two_sd = TRUE)[[2]][-1], tolerance = 0.01)
   })
 
 
@@ -109,7 +109,7 @@ if (require("testthat") && require("effectsize")) {
 
     testthat::expect_equal(standardize_parameters(m_aov),
                            standardize_parameters(m_lm),
-                           check.attributes = FALSE)
+                           ignore_attr = TRUE)
   })
 
 
@@ -232,18 +232,18 @@ if (require("testthat") && require("effectsize")) {
       testthat::expect_equal(
         suppressWarnings(standardize_parameters(model, method = "refit")$Std_Median[1:4]),
         c(0.065, -0.094, -0.100, 0.862),
-        tol = 0.01
+        tolerance = 0.01
       )
 
       testthat::expect_equal(
[…]
