### 2020-11-11 #2815 @trashbirdecology
Dearest Whedon,

@mattansb has an active issue in their repository with a few things to
resolve before completing the joss checklist
