### 2020-11-11 easystats/effectsize#196 @mattansb
Hey @trashbirdecology - sorry, but I'm not sure if you're done listing these or not?
(from you recent commit on the review issue, sounds like you're waiting for me?)
