### 2020-11-11 easystats/effectsize#196 @trashbirdecology
oh, sorry about that @mattansb. I would like to go back through after these issues are resolved (or not).
