### 2020-11-11 easystats/effectsize#196 @mattansb
Okay great - then I'll start working on these and let you know when I'm done addressing these.
