+++ paper/paper.bib
@@ -87,3 +87,59 @@
   note = {R package version 1.3.5},
   url = {https://CRAN.R-project.org/package=gtsummary},
 }
+
+@article{luedecke2019insight,
+  title = {insight: A Unified Interface to Access Information from Model Objects in R.},
+  volume = {4},
+  doi = {10.21105/joss.01412},
+  number = {38},
+  journal = {Journal of Open Source Software},
+  author = {Daniel Lüdecke and Philip Waggoner and Dominique Makowski},
+  year = {2019},
+  pages = {1412},
+}
+
+@manual{behrendt2014lmbeta,
+  title = {lm.beta: Add Standardized Regression Coefficients to lm-Objects},
+  author = {Stefan Behrendt},
+  year = {2014},
+  note = {R package version 1.5-1},
+  url = {https://CRAN.R-project.org/package=lm.beta},
+}
+
+@manual{buchanan2019MOTE,
+  title = {{MOTE: Measure of the Effect}: Package to assist in effect size calculations and their confidence intervals},
+  author = {Erin M. Buchanan and Amber Gillenwaters and John E. Scofield and K.D. Valentine},
+  year = {2019},
+  note = {R package version 1.0.2},
+  url = {http://github.com/doomlab/MOTE},
+}
+
+@manual{kelley2020MBESS,
+  title = {MBESS: The MBESS R Package},
+  author = {Ken Kelley},
+  year = {2020},
+  note = {R package version 4.8.0},
+  url = {https://CRAN.R-project.org/package=MBESS},
+}
+
+@book{cohen1988statistical,
+  title={Statistical power analysis for the behavioral sciences, 2nd edn. {\'A}/L},
+  author={Cohen, J},
+  year={1988},
+  publisher={Erbaum Press, Hillsdale, NJ, USA}
+}
+
+@article{hedges1985statistical,
+  title={Statistical Methods for Meta-Analysis (Orlando, FL: Academic)},
+  author={Hedges, L and Olkin, I},
+  journal={HedgesStatistical Methods for Meta-Analysis1985},
+  year={1985}
+}
+
+@book{hedges2014statistical,
+  title={Statistical methods for meta-analysis},
+  author={Hedges, Larry V and Olkin, Ingram},
+  year={2014},
+  publisher={Academic press}
+}
