### 2020-11-12 easystats/effectsize#196 @mattansb
I think I've addresses all the issues you've raised. You can find them in the master branch (or, if you'd prefer a comparison, you can take a look at this PR #202 ). Below are some direct comments to some of your points / questions.

Please let me know if anything is missing or unclear.

Thanks @trashbirdecology !

## Required fixes for JOSS review

> question/minor: should it be noted that it's currently available via CRAN? (not sure JOSS requirements or recommendations on this)

This wasn't required for our previous submissions, so if that's any indication, I think this is not necessary.

> "returns an updated model," the function doesn't return an updated model, it provides updated coefficients, including Classes ‘effectsize_table’, ‘see_effectsize_table’, ‘effectsize_std_params’. I.e., the output of  `standardize_parameters()` does not == the output of `lm()`

I have made the distinction between `standardize()` (which returns an updated model) and `standardize_parameters()` (which returns a table of parameters) more explicit in the example. Please let me know if this is now clearer.

> suggest: "from a provided model" -- briefly list the available model types which `standardize_paramters` can handle

I've added a reference to the `insight` package, which is the backbone of the easystats ecosystem.

> Dependency warnings on `devtools::document()`

All easystats packages are designed to be light-weight on dependencies, so when the packages are not available these fails are the intended behavior - asking users to only install any additional packages as they are needed. When running the tests with all the suggested packages installed, all tests pass (as is done in CRAN-checks).

> FAILURE `devtools::check()` on vignette build @mattansb

This has been fixed - thanks for catching this!

## JLB suggestions (not required for JOSS review)

> feature request: as a potential user I would find it useful to a  quick-reference table in the paper and/or readme that allows me to easily identify what my standardization, conversion, and interpretation options are.

I think these are too many to list concisely in a single table - but I've added a hyper link to https://easystats.github.io/effectsize/reference/index.html in both the paper and the readme ("a full list of features and functions").

> Rename primary branch from 'master' to 'main' (or something else)

Great suggestion! I will definitely do this - when the review is completed, as the JOSS review is currently linked to the "master" branch.

## JLB to do list

> Check singularity warning when running following example in paper. I haven't yet looked into this,  just noting here to remember to do so.

Oops, forgot to add a line that filters out any chickens with incomplete data (never thought I'd write that sentence...). Fixed.

> Tests won't run using `testthat::test_check()` (@mattansb if you can recommend a quick fix or simply point out the reason test_check() doesn't work but test() does, I would be grateful

Seems like `test_check()` shouldn't be called directly? [As per Hadely](https://github.com/r-lib/testthat/issues/205#issuecomment-69342965)
