### 2020-11-12 easystats/effectsize#196 @trashbirdecology
Thanks for responding to some of my questions and reference to test_check note!
The only thing I haven't checked off is the first reference to @cohen1998statistical appears at the second instance of Cohen's d.
> " such as Cohen's d "

This is minor, so I will go ahead and resolve this issue.
