### 2020-11-12 #2815 @mattansb
Usually "sum of squares", but the context here is of several of these (with multiple tests statistics, multiple effect sizes, and so also many sums of squares).
