### 2020-11-16 fc871f6
exclude is_censored explicitely, due to revisions in model_info()
