+++ R/standardize.models.R
@@ -310,7 +310,7 @@ standardize.wbgee <- standardize.wbm
 #' @keywords internal
 .no_response_standardize <- function(info) {
   # check if model has a response variable that should not be standardized.
-  !info$is_linear | info$family == "inverse.gaussian"
+  !info$is_linear | info$is_censored | info$family == "inverse.gaussian"
 
   ## TODO alternative would be to keep the below line for checking if no std possible
   ##      and then treat response for "Gamma()" or "inverse.gaussian" similar to log-terms
