### 2020-11-24 6e54970
abracadabob interpret is now a method
