+++ man/interpret.Rd
@@ -4,14 +4,12 @@
 \alias{interpret}
 \title{Generic function for interpretation}
 \usage{
-interpret(x, rules, name = attr(rules, "rule_name"))
+interpret(x, ...)
 }
 \arguments{
 \item{x}{Vector of value break points (edges defining categories).}
 
 \item{rules}{Set of \code{\link[=rules]{rules()}}.}
-
-\item{name}{Name of the set of rules (stored as a 'rule_name' attribute).}
 }
 \description{
 Interpret a value based on a set of rules. See \code{\link[=rules]{rules()}}.
