+++ NAMESPACE
@@ -15,6 +15,7 @@ S3method(equivalence_test,effectsize_table)
 S3method(eta_squared_posterior,brmsfit)
 S3method(eta_squared_posterior,stanreg)
 S3method(interpret,lavaan)
+S3method(interpret,numeric)
 S3method(interpret,performance_lavaan)
 S3method(interpret_parameters,lm)
 S3method(normalize,data.frame)
