+++ man/interpret.lavaan.Rd
@@ -1,9 +1,12 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/interpret.lavaan.R
-\name{interpret.lavaan}
+\name{interpret.performance_lavaan}
+\alias{interpret.performance_lavaan}
 \alias{interpret.lavaan}
 \title{Interpretation for lavaan objects}
 \usage{
+\method{interpret}{performance_lavaan}(x, ...)
+
 \method{interpret}{lavaan}(x, ...)
 }
 \arguments{
@@ -23,6 +26,6 @@ if (require("lavaan")) {
                  dem60 =~ y1 + y2 + y3
                  dem60 ~ ind60 "
   model <- lavaan::sem(structure, data = PoliticalDemocracy)
-  interpret(model)
+  # interpret(model)  # Not working until new performance is up
 }
 }
