+++ _pkgdown.yml
@@ -46,6 +46,7 @@ reference:
   contents:
   - equivalence_test.effectsize_table
   - interpret
+  - interpret.performance_lavaan
   - rules
   - interpret_d
   - interpret_r
