+++ R/interpret_cfa_fit.R
@@ -2,7 +2,7 @@
 #'
 #' Interpretation of indices of fit found in confirmatory analysis or structural equation modelling, such as RMSEA, CFI, NFI, IFI, etc.
 #'
-#' @param x vector of values.
+#' @param x vector of values, or an object of class `lavaan`.
 #' @param rules Can be `"default"` or custom set of [rules()].
 #'
 #' @inherit performance::model_performance.lavaan details
@@ -40,6 +40,15 @@
 #' interpret_rfi(c(.5, .99))
 #' interpret_ifi(c(.5, .99))
 #' interpret_pnfi(c(.5, .99))
+#'
+#' # Structural Equation Models (SEM)
+#' if (require("lavaan")) {
+#'   structure <- " ind60 =~ x1 + x2 + x3
+#'                  dem60 =~ y1 + y2 + y3
+#'                  dem60 ~ ind60 "
+#'   model <- lavaan::sem(structure, data = PoliticalDemocracy)
+#'   # interpret(model)  # Not working until new performance is up
+#' }
 #'
 #' @references
 #' - Awang, Z. (2012). A handbook on SEM. Structural equation modeling.
@@ -176,4 +185,111 @@ interpret_pnfi <- function(x, rules = "default") {
   )
 
   interpret(x, rules)
+}
+
+
+# lavaan ------------------------------------------------------------------
+
+#' @rdname interpret_gfi
+#' @export
+interpret.lavaan <- function(x, ...) {
+  interpret(performance::model_performance(x, ...), ...)
+}
+
+#' @rdname interpret_gfi
+#' @export
+interpret.performance_lavaan <- function(x, ...){
+  table <- data.frame(Name = c(), Value = c(), Interpretation = c(), Threshold = c())
+
+  # GFI
+  if ("GFI" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "GFI",
+                              Value = x$GFI,
+                              Interpretation = interpret_gfi(x$GFI),
+                              Threshold = 0.95))
+  }
+
+  # AGFI
+  if ("AGFI" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "AGFI",
+                              Value = x$AGFI,
+                              Interpretation = interpret_agfi(x$AGFI),
+                              Threshold = 0.90))
+  }
+
+  # NFI
+  if ("NFI" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "NFI",
+                              Value = x$NFI,
+                              Interpretation = interpret_nfi(x$NFI, rules = "byrne1994"),
+                              Threshold = 0.90))
+  }
+
+  # NNFI
+  if ("NNFI" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "NNFI",
+                              Value = x$NNFI,
+                              Interpretation = interpret_nnfi(x$NNFI, rules = "byrne1994"),
+                              Threshold = 0.90))
+  }
+
+  # CFI
+  if ("CFI" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "CFI",
+                              Value = x$CFI,
+                              Interpretation = interpret_cfi(x$CFI),
+                              Threshold = 0.90))
+  }
+
+  # RMSEA
+  if ("RMSEA" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "RMSEA",
+                              Value = x$RMSEA,
+                              Interpretation = interpret_rmsea(x$RMSEA),
+                              Threshold = 0.05))
+  }
+
+  # SRMR
+  if ("SRMR" %in% names(x)) {
+    table <- rbind(table,
+                   data.frame(Name = "SRMR",
+                              Value = x$SRMR,
[…]
