### 2020-11-26 a2ae2ae
Use export_table
https://github.com/easystats/easystats/issues/83
