+++ R/print.rules.R
@@ -9,7 +9,7 @@ print.rules <- function(x, ...){
     df <- data.frame(setNames(as.list(x$values), x$labels), check.names = FALSE)
     df <- cbind("Label " = "Value ", df)
     insight::print_color(paste0("# Reference values (",name,")\n\n"), "blue")
-    cat(insight::format_table(df))
+    cat(insight::export_table(df))
   } else{
     if (isTRUE(attr(x, "right"))) {
       gLeft <- " <= "
