### 2020-11-26 a2ab3d8
update rules example
