+++ R/interpret.R
@@ -20,7 +20,7 @@
 #' @seealso interpret
 #'
 #' @examples
-#' rules(c(0.05), c("significant", "not significant"))
+#' rules(c(0.05), c("significant", "not significant"), right = FALSE)
 #' rules(c(0.2, 0.5, 0.8), c("small", "medium", "large"))
 #' rules(c("small" = 0.2, "medium" = 0.5), name = "Cohen's Rules")
 #'
