+++ man/rules.Rd
@@ -29,7 +29,7 @@ or in the interval to the left (higher values).}
 Create a container for interpretation rules of thumb. Usually used in conjunction with \link{interpret}.
 }
 \examples{
-rules(c(0.05), c("significant", "not significant"))
+rules(c(0.05), c("significant", "not significant"), right = FALSE)
 rules(c(0.2, 0.5, 0.8), c("small", "medium", "large"))
 rules(c("small" = 0.2, "medium" = 0.5), name = "Cohen's Rules")
 
