+++ R/convert_tFz_to_r.R
@@ -33,8 +33,9 @@
 #' \deqn{d = 2 * z / \sqrt{N}}
 #'
 #' The resulting `d` effect size is an *approximation* to Cohen's *d*, and
-#' assumes equal group sizes. When possible, it is advised to directly estimate
-#' Cohen's *d*, with [cohens_d()], [emmeans::eff_size()], or similar functions.
+#' assumes two equal group sizes. When possible, it is advised to directly
+#' estimate Cohen's *d*, with [cohens_d()], [emmeans::eff_size()], or similar
+#' functions.
 #'
 #' @inheritSection cohens_d Confidence Intervals
 #'
