+++ man/t_to_r.Rd
@@ -60,8 +60,9 @@ These functions use the following formulae to approximate \emph{r} and \emph{d}:
 \deqn{d = 2 * z / \sqrt{N}}
 
 The resulting \code{d} effect size is an \emph{approximation} to Cohen's \emph{d}, and
-assumes equal group sizes. When possible, it is advised to directly estimate
-Cohen's \emph{d}, with \code{\link[=cohens_d]{cohens_d()}}, \code{\link[emmeans:eff_size]{emmeans::eff_size()}}, or similar functions.
+assumes two equal group sizes. When possible, it is advised to directly
+estimate Cohen's \emph{d}, with \code{\link[=cohens_d]{cohens_d()}}, \code{\link[emmeans:eff_size]{emmeans::eff_size()}}, or similar
+functions.
 }
 \section{Confidence Intervals}{
 Confidence intervals are estimated using the Noncentrality parameter method;
