### 2020-12-01 b5d1a55
#215
Fix issue introduced with complete.cases
