### 2020-12-01 7d3d0f7
Update README.Rmd
https://github.com/easystats/easystats/issues/85
