### 2020-12-02 e79fed2
fix CI for non-pooled sd
invoke @IndrajeetPatil to find where I fucked up.
