+++ tests/testthat/test-standardized_differences.R
@@ -45,8 +45,8 @@ if (require("testthat") && require("effectsize")) {
     x <- cohens_d(wt ~ am, data = mtcars, pooled_sd = FALSE)
     testthat::expect_equal(colnames(x)[1], "Cohens_d")
     testthat::expect_equal(x[[1]], 1.934, tolerance = 0.001)
-    testthat::expect_equal(x$CI_low, 1.102, tolerance = 0.001)
-    testthat::expect_equal(x$CI_high, 2.829, tolerance = 0.001)
+    testthat::expect_equal(x$CI_low, 1.098798, tolerance = 0.001)
+    testthat::expect_equal(x$CI_high, 2.833495, tolerance = 0.001)
   })
 
   test_that("hedges_g", {
