### 2020-12-02 45e9b86
add `effectsize(type = ,...)`
#214
