+++ R/effectsize.R
@@ -4,20 +4,21 @@
 #' input model. See details.
 #'
 #' @param model An object of class `htest`, or a statistical model. See details.
+#' @param type The effect size of interest. See details.
 #' @param ... Arguments passed to or from other methods. See details.
 #'
 #' @details
 #'
 #' - For an object of class `htest`:
 #'   - A **t-test** returns *Cohen's d* via [t_to_d()].
-#'   - A **correlation test** returns *r*. See [t_to_r()].
-#'   - A **Chi-squared test** returns *Cramer's V* via [cramers_v()].
-#'   - A **One-way ANOVA test** returns *Eta squared* via [F_to_eta2()], but can be changes via an `es` argument.
+#'   - A **correlation test** returns *r*.
+#'   - A **Chi-squared test**, depending on `type`: `"cramers_v"` (default), `"phi"` or `"cohens_w"`, `"oddsratio"`, or `"riskratio"`.
+#'   - A **One-way ANOVA test**, depending on `type`: `"eta"` (default), `"omega"` or `"epsilon"` -squared, `"f"`, or `"f2"`.
 #' - For an object of class `BFBayesFactor`, using [bayestestR::describe_posterior()],
 #'   - A **t-test** returns *Cohen's d*.
-#'   - A **correlation test** returns *r*..
-#'   - A **contingency table test** returns *Cramer's V*.
-#' - Objects of class `anova`, `aov`, or `aovlist` are passed to [eta_squared()].
+#'   - A **correlation test** returns *r*.
+#'   - A **contingency table test**, depending on `type`: `"cramers_v"` (default), `"phi"` or `"cohens_w"`, `"oddsratio"`, or `"riskratio"`.
+#' - Objects of class `anova`, `aov`, or `aovlist`, depending on `type`: `"eta"` (default), `"omega"` or `"epsilon"` -squared, `"f"`, or `"f2"`.
 #' - Other objects are passed to [standardize_parameters()].
 #'
 #' **For statistical models it is recommended to directly use the listed
@@ -32,12 +33,14 @@
 #' contingency_table <- as.table(rbind(c(762, 327, 468), c(484, 239, 477), c(484, 239, 477)))
 #' Xsq <- chisq.test(contingency_table)
 #' effectsize(Xsq)
+#' effectsize(Xsq, type = "phi")
 #'
 #' Ts <- t.test(1:10, y = c(7:20))
 #' effectsize(Ts)
 #'
-#' Aov <- oneway.test(extra ~ group, data = sleep)
+#' Aov <- oneway.test(extra ~ group, data = sleep, var.equal = TRUE)
 #' effectsize(Aov)
+#' effectsize(Aov, type = "omega")
 #'
 #'
 #' ## Bayesian Hypothesis Testing
@@ -52,6 +55,7 @@
 #'   data(raceDolls)
 #'   bf3 <- contingencyTableBF(raceDolls, sampleType = "poisson", fixedMargin = "cols")
 #'   effectsize(bf3, test = NULL)
+#'   effectsize(bf3, type = "oddsratio", test = NULL)
 #' }
 #'
 #'
@@ -62,6 +66,7 @@
 #'
 #' anova_table <- anova(fit)
 #' effectsize(anova_table)
+#' effectsize(anova_table, type = "epsilon")
 #'
 #' @export
 effectsize <- function(model, ...) {
@@ -69,7 +74,8 @@ effectsize <- function(model, ...) {
 }
 
 #' @export
-effectsize.htest <- function(model, ...) {
+#' @rdname effectsize
+effectsize.htest <- function(model, type = NULL, ...) {
   if (grepl("t-test", model$method)) {
     # message("Using t_to_d().")
[…]
