+++ tests/testthat/test-effectsize.R
@@ -34,6 +34,20 @@ if (require("testthat") && require("effectsize")) {
       effectsize(Xsq3)$Cramers_v,
       cramers_v(table(mtcars$cyl))$Cramers_v
     )
+
+    # types
+    testthat::expect_equal(effectsize(Xsq1, type = "phi"),
+                           phi(contingency_table))
+    testthat::expect_equal(effectsize(Xsq1, type = "w"),
+                           cohens_w(contingency_table))
+
+    testthat::expect_error(effectsize(Xsq1, type = "riskratio"))
+    contingency_table22 <- contingency_table[1:2,1:2]
+    Xsq4 <- chisq.test(contingency_table22)
+    testthat::expect_equal(effectsize(Xsq4, type = "oddsratio"),
+                           oddsratio(contingency_table22))
+    testthat::expect_equal(effectsize(Xsq4, type = "riskratio"),
+                           riskratio(contingency_table22))
   })
 
   test_that("cor.test", {
@@ -52,12 +66,30 @@ if (require("testthat") && require("effectsize")) {
   })
 
 
+  test_that("one way", {
+    onew <- oneway.test(mpg ~ cyl, mtcars)
+    testthat::expect_warning(effectsize(onew))
+
+
+    onew <- oneway.test(mpg ~ cyl, mtcars, var.equal = TRUE)
+    m <- aov(mpg ~ cyl, mtcars)
+
+    testthat::expect_equal(eta_squared(m, partial = FALSE)[,-1], effectsize(onew),
+                           tolerance = 0.03, ignore_attr = TRUE)
+    testthat::expect_equal(omega_squared(m, partial = FALSE)[,-1], effectsize(onew, type = "omega"),
+                           tolerance = 0.03, ignore_attr = TRUE)
+    testthat::expect_equal(cohens_f(m, partial = FALSE)[,-1], effectsize(onew, type = "f"),
+                           tolerance = 0.03, ignore_attr = TRUE)
+  })
+
+
   # aov ---------------------------------------------------------------------
   test_that("aov", {
     data <- iris
     data$Cat1 <- rep(c("A", "B"), length.out = nrow(data))
     model <- aov(Sepal.Length ~ Species * Cat1, data = data)
     testthat::expect_equal(effectsize(model), eta_squared(model))
+    testthat::expect_equal(effectsize(model, type = "omega"), omega_squared(model))
   })
 
 
@@ -69,6 +101,7 @@ if (require("testthat") && require("effectsize")) {
       data(raceDolls)
       bf1 <- contingencyTableBF(raceDolls, sampleType = "poisson", fixedMargin = "cols")
       testthat::expect_equal(effectsize(bf1, test = NULL)[[2]], 0.164, tolerance = 0.01)
+      testthat::expect_equal(effectsize(bf1, test = NULL, type = "OR")[[2]], 0.498, tolerance = 0.01)
 
       bf2 <- ttestBF(mtcars$mpg[mtcars$am == 1], mtcars$mpg[mtcars$am == 0])
       testthat::expect_equal(effectsize(bf2, test = NULL)[[2]], 1.30, tolerance = 0.01)
