+++ NEWS.md
@@ -1,4 +1,4 @@
-# effectsize 0.4.0.001
+# effectsize 0.4.1
 
 ## Breaking Changes
 
@@ -6,6 +6,7 @@
 
 ## New features
 
+- `effectsize()` gains `type = ` argument for specifying which effect size to return.
 - `eta_squared_posterior()` can return a generalized Eta squared.
 - `oddsratio()` and `riskratio()` functions for 2-by-2 contingency tables.
 - `standardize()` gains support for `mediation::mediate()` models.
