### 2020-12-02 605d15d
Use testthat::skip_if_not_installed
