+++ tests/testthat/test-eta_squared_etc.R
@@ -64,31 +64,31 @@ if (require("testthat") && require("effectsize")) {
     testthat::expect_true(all(c("Group", "Parameter") %in% colnames(res)))
 
 
-    if (require("afex")) {
-      # non-partial Eta2 should be the same for aov and aovlist
-      data(obk.long, package = "afex")
-      model <- afex::aov_car(value ~ treatment * gender + Error(id/(phase*hour)),
-                             data = obk.long, observed = "gender",
-                             include_aov = TRUE)
 
-      model2 <- aov(value ~ treatment * gender * phase * hour,
-                    data = model$data$long,
-                    contrasts = list(
-                      treatment = contr.sum,
-                      gender = contr.sum,
-                      phase = contr.sum,
-                      hour = contr.sum
-                    ))
+    testthat::skip_if_not_installed("afex")
+    # non-partial Eta2 should be the same for aov and aovlist
+    data(obk.long, package = "afex")
+    model <- afex::aov_car(value ~ treatment * gender + Error(id/(phase*hour)),
+                           data = obk.long, observed = "gender",
+                           include_aov = TRUE)
 
-      a1 <- eta_squared(model2, partial = F)
-      a2 <- eta_squared(model$aov, partial = F)
+    model2 <- aov(value ~ treatment * gender * phase * hour,
+                  data = model$data$long,
+                  contrasts = list(
+                    treatment = contr.sum,
+                    gender = contr.sum,
+                    phase = contr.sum,
+                    hour = contr.sum
+                  ))
 
-      rownames(a1) <- a1$Parameter
-      rownames(a2) <- a2$Parameter
+    a1 <- eta_squared(model2, partial = F)
+    a2 <- eta_squared(model$aov, partial = F)
 
-      testthat::expect_equal(a1[a1$Parameter, "Eta2"],
-                             a2[a1$Parameter, "Eta2"])
-    }
+    rownames(a1) <- a1$Parameter
+    rownames(a2) <- a2$Parameter
+
+    testthat::expect_equal(a1[a1$Parameter, "Eta2"],
+                           a2[a1$Parameter, "Eta2"])
   })
 
   # mlm / anova table -------------------------------------------------------
@@ -131,112 +131,113 @@ if (require("testthat") && require("effectsize")) {
     rownames(fsD) <- rownames(fs) <- 1
     testthat::expect_equal(fsD, fs, tolerance = 0.01)
 
-    if (require("performance")) {
-      fsD <- cohens_f_squared(m1, model2 = m2)
 
-      R2_1 <- performance::r2(m1)[[1]]
-      R2_2 <- performance::r2(m2)[[1]]
-
-      testthat::expect_equal(fsD$Cohens_f2_partial,
-                             unname((R2_2 - R2_1) / (1 - R2_2)))
-    }
+    testthat::skip_if_not_installed("performance")
+    fsD <- cohens_f_squared(m1, model2 = m2)
+    R2_1 <- performance::r2(m1)[[1]]
+    R2_2 <- performance::r2(m2)[[1]]
+    testthat::expect_equal(fsD$Cohens_f2_partial,
+                           unname((R2_2 - R2_1) / (1 - R2_2)))
   })
 
   # generalized Eta -------------------------------------------------------------
-  if (require("car") && require("afex")) {
-    test_that("generalized | between", {
-      data(obk.long, package = "afex")
-      m <- suppressWarnings(
-        afex::aov_car(value ~ treatment * gender + Error(id),
-                      data = obk.long, observed = "gender",
-                      include_aov = TRUE)
-      )
+  test_that("generalized | between", {
[…]
