+++ R/print.effectsize_table.R
@@ -1,5 +1,5 @@
 #' @export
-print.effectsize_table <- function(x, digits = 2, ...){
+print.effectsize_table <- function(x, digits = 2, ...) {
   x_orig <- x
 
   i <- is_effectsize_name(colnames(x))
@@ -9,10 +9,11 @@ print.effectsize_table <- function(x, digits = 2, ...){
     ci_level <- x$CI[1]
 
     x$CI <- insight::format_ci(x$CI_low,
-                               x$CI_high,
-                               ci = NULL,
-                               digits = digits,
-                               width = "auto")
+      x$CI_high,
+      ci = NULL,
+      digits = digits,
+      width = "auto"
+    )
 
     colnames(x)[colnames(x) == "CI"] <- sprintf("%g%% CI", round(ci_level * 100, digits = digits))
 
@@ -29,4 +30,3 @@ print.effectsize_table <- function(x, digits = 2, ...){
 
   invisible(x_orig)
 }
-
