+++ R/convert_odds_to_probs.R
@@ -18,7 +18,6 @@
 #'
 #' probs_to_odds(0.95)
 #' probs_to_odds(0.95, log = TRUE)
-#'
 #' @export
 odds_to_probs <- function(odds, log = FALSE, ...) {
   UseMethod("odds_to_probs")
