+++ NEWS.md
@@ -12,6 +12,10 @@
 - `standardize()` gains support for `mediation::mediate()` models.
 - `eta_squared()` family available for `manova` objects.
 
+## Changes
+
+- `eta_squared()` family of functions returns non-partial effect size for one-way between subjects design (#180).
+
 ## Bug fixes
 
 Fixed width of CI for Cohen's d and Hedge's g when using *non*-pooled SD.
