### 2020-12-03 a8688f8
Update eta_squared.R
being stylish
