+++ pkgdown.yml
@@ -9,5 +9,5 @@ articles:
   interpret: interpret.html
   standardize_data: standardize_data.html
   standardize_parameters: standardize_parameters.html
-last_built: 2020-12-02T09:25Z
+last_built: 2020-12-03T13:40Z
 
