### 2020-12-03 #2815 @jkarreth
I will submit my initial review around December 10. Apologies again for taking so long to get started (I had committed to this review saying that I probably wouldn't get to it until the end of the current JOSS review time window, and sadly that turned out to be true.)

Major thanks to @trashbirdecology for her thorough and timely review - that'll likely leave very few issues for me to find 👍
