### 2020-12-04 2069bbd
clean up printing code
