+++ R/print.effectsize_difference.R
@@ -6,9 +6,10 @@
 #' @param ... Not used.
 print.effectsize_difference <- function(x, digits = 2, append_CL = FALSE, ...) {
   x_orig <- x
-  print.effectsize_table(x, digits = digits)
 
-  ## Add note
+  footer <- caption <- NULL
+
+  ## Add footer
   # note <- NULL
   # if (any(colnames(x) %in% c("Cohens_d", "Hedges_g"))) {
   #   note <- c(note, ifelse(attr(x, "pooled_sd"), "Pooled SD", "Unpooled SD"))
@@ -18,14 +19,17 @@ print.effectsize_difference <- function(x, digits = 2, append_CL = FALSE, ...) {
   # }
   # if (!is.null(note)) cat(c("\n[",paste0(note, collapse = "; ")), "]\n")
 
-  ## Common lang
   if (append_CL && any(colnames(x) %in% c("Cohens_d", "Hedges_g"))) {
+    # Common lang
     cl <- d_to_common_language(x[[any(colnames(x) %in% c("Cohens_d", "Hedges_g"))]])
-    cl <- sapply(cl, function(ff) sprintf("%g%% CI", round(ff * 100, digits = digits)))
-    cl <- paste(paste0("* ", names(cl), ": ", cl), collapse = "\n")
-    cat("\n")
-    insight::print_color(cl, "cyan")
+    cl <- sapply(cl, function(ff) sprintf("%g%%", round(ff * 100, digits = digits)))
+    cl <- paste(paste0("\n* ", names(cl), ": ", cl), collapse = "")
+
+    footer <- c(cl, "cyan")
   }
+
+  x <- .print_effectsize_table(x, digits = digits)
+  cat(insight::export_table(x, digits = digits, caption = caption, footer = footer))
 
   invisible(x_orig)
 }
