### 2020-12-04 bff3eef
Add std-diff infor to printing
