### 2020-12-05 c39f329
"fix" d and g (with corrections) #222
