+++ R/cohens_d.R
@@ -1,24 +1,33 @@
 #' Effect size for differences
 #'
-#' Compute different indices of effect size. For very small sample sizes (n <
-#' 20) Hedges' g is considered as less biased than Cohen's d. For sample sizes >
-#' 20, the results for both statistics are roughly equivalent.
+#' Compute effect size indices for standardized differences: Cohen's *d*,
+#' Hedges' *g* and Glass’s *delta*. (This function returns the **population**
+#' estimate.)
 #' \cr\cr
-#' The Glass’s delta is appropriate if standard deviations are significantly
-#' different between groups, as it uses only the *second* group's standard
-#' deviation.
+#' Both Cohen's *d* and Hedges' *g* are the estimated the standardized
+#' difference between the means of two populations. Hedges' *g* provides a bias
+#' correction to Cohen's *d* for small sample sizes. For sample sizes > 20, the
+#' results for both statistics are roughly equivalent. Glass’s *delta* is
+#' appropriate when the standard deviations are significantly different between
+#' the populations, as it uses only the *second* group's standard deviation.
 #'
 #' @param x A formula, a numeric vector, or a character name of one in `data`.
 #'   (For `print()` the result of one of the standardized difference functions.)
 #' @param y A numeric vector, a grouping (character / factor) vector, a or a
 #'   character  name of one in `data`. Ignored if `x` is a formula.
 #' @param data An optional data frame containing the variables.
-#' @param correction If `TRUE`, applies a correction to make it less biased for
-#'   small samples (McGrath & Meyer, 2006).
+#' @param correction Type of small sample bias correction to apply to produce
+#'   Hedges' *g*. Can be `1` for Hedges and Olkin's original correction
+#'   (default) or `2` for Hunter and Schmidt's correction (see McGrath & Meyer,
+#'   2006).
 #' @param pooled_sd If `TRUE` (default), a [sd_pooled()] is used (assuming equal
 #'   variance). Else the mean SD from both groups is used instead.
 #' @param paired If `TRUE`, the values of `x` and `y` are considered as paired.
 #' @inheritParams chisq_to_phi
+#'
+#' @note The indices here give the population estimated standardized difference.
+#'   Some statistical packages give the sample estimate instead (without
+#'   applying Bessel's correction).
 #'
 #' @details
 #'
@@ -46,24 +55,29 @@
 #' print(cohens_d(mpg ~ am, data = mtcars), append_CL = TRUE)
 #' @references
 #' - Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.
-#' - McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
 #' - Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.
+#' - Hunter, J. E., & Schmidt, F. L. (2004). Methods of meta-analysis: Correcting error and bias in research findings. Sage.
+#' - McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
 #'
 #' @importFrom stats var model.frame
 #' @export
 cohens_d <- function(x,
                      y = NULL,
                      data = NULL,
-                     correction = FALSE,
                      pooled_sd = TRUE,
                      paired = FALSE,
-                     ci = 0.95) {
+                     ci = 0.95,
+                     correction) {
+  if (!missing(correction)) {
[…]
