+++ tests/testthat/test-printing.R
@@ -13,14 +13,17 @@ if (require("testthat") && require("effectsize")) {
   test_that("effectsize difference", {
     d <- cohens_d(1:3, c(1,1:3))
     testthat::expect_output(print(d), regexp = "Cohen")
-    testthat::expect_output(print(d), regexp = "Pooled")
+    testthat::expect_output(print(d), regexp = "pooled")
     testthat::expect_output(print(d, append_CL = TRUE), regexp = "U3")
 
-    d <- cohens_d(1:3, c(1,1:3), correction = TRUE)
-    testthat::expect_output(print(d), regexp = "Bias-corrected")
-
     d <- cohens_d(1:3, c(1,1:3), pooled_sd = FALSE)
-    testthat::expect_output(print(d), regexp = "Unpooled")
+    testthat::expect_output(print(d), regexp = "un-pooled")
+
+    d <- hedges_g(1:3, c(1,1:3), correction = 1)
+    testthat::expect_output(print(d), regexp = "Hedges and Olkin")
+
+    d <- hedges_g(1:3, c(1,1:3), correction = 2)
+    testthat::expect_output(print(d), regexp = "Hunter and Schmidt")
   })
 
   test_that("equivalence test effectsize", {
