### 2020-12-05 edc3d86
fix printing of CL for std-d
