### 2020-12-05 d8931ab
Explain paired Cohen's d
