### 2020-12-05 d1591b0
Bump bayestestR version
