+++ DESCRIPTION
@@ -38,7 +38,7 @@ Description: Provide utilities to work with indices of effect size and standardi
 Depends:
     R (>= 3.5)
 Imports:
-    bayestestR (>= 0.7.5),
+    bayestestR (>= 0.8.0),
     insight (>= 0.11.0),
     parameters (>= 0.10.0),
     stats,
