### 2020-12-06 52a49be
Update DESCRIPTION
Dependence on R version '3.5.2' not with patchlevel 0
