### 2020-12-06 beba581
hard code standardize.data.frame, NAs / append
