+++ tests/testthat/test-standardize-data.R
@@ -36,33 +36,31 @@ if (require("testthat") && require("effectsize")) {
 
   test_that("standardize.data.frame, NAs", {
     data(iris)
-    set.seed(123)
-    iris$Sepal.Width[sample(1:150, 10)] <- NA
-    iris$Sepal.Length[sample(1:150, 10)] <- NA
+    iris$Sepal.Width[c(148, 65, 33, 58, 54, 93, 114, 72, 32, 23)] <- NA
+    iris$Sepal.Length[c(11, 30, 141, 146, 13, 149, 6, 8, 48, 101)] <- NA
 
     x <- standardize(iris)
-    testthat::expect_equal(head(x$Sepal.Length), c(-0.9163, -1.1588, -1.4013, -1.5226, -1.0376, -0.5526), tolerance = 0.01)
-    testthat::expect_equal(head(x$Sepal.Width), c(0.9965, -0.1377, 0.316, 0.0891, 1.2233, 1.9038), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Length), c(-0.9163, -1.1588, -1.4013, -1.5226, -1.0376, NA), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Width), c(1.0237, -0.151, 0.3189, 0.0839, 1.2586, 1.9635), tolerance = 0.01)
     testthat::expect_equal(mean(x$Sepal.Length), as.numeric(NA))
 
     x <- standardize(iris, two_sd = TRUE)
-    testthat::expect_equal(head(x$Sepal.Length), c(-0.4582, -0.5794, -0.7007, -0.7613, -0.5188, -0.2763), tolerance = 0.01)
-    testthat::expect_equal(head(x$Sepal.Width), c(0.4982, -0.0689, 0.158, 0.0446, 0.6116, 0.9519), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Length), c(-0.4603, -0.5811, -0.7019, -0.7623, -0.5207, NA), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Width), c(0.5118, -0.0755, 0.1594, 0.042, 0.6293, 0.9817), tolerance = 0.01)
     testthat::expect_equal(mean(x$Sepal.Length), as.numeric(NA))
 
     testthat::skip_if_not_installed("dplyr")
     x <- standardize(dplyr::group_by(iris, .data$Species))
-    testthat::expect_equal(head(x$Sepal.Length), c(0.2086, -0.3681, -0.9447, -1.233, -0.0797, 1.0735), tolerance = 0.01)
-    testthat::expect_equal(head(x$Sepal.Width), c(0.1441, -1.1586, -0.6375, -0.8981, 0.4047, 1.1863), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Length), c(0.2547, -0.3057, -0.8661, -1.1463, -0.0255, NA), tolerance = 0.01)
+    testthat::expect_equal(head(x$Sepal.Width), c(0.2369, -1.0887, -0.5584, -0.8235, 0.502, 1.2974), tolerance = 0.01)
     testthat::expect_equal(mean(x$Sepal.Length), as.numeric(NA))
   })
 
 
   test_that("standardize.data.frame, apend", {
     data(iris)
-    set.seed(123)
-    iris$Sepal.Width[sample(1:150, 10)] <- NA
-    iris$Sepal.Length[sample(1:150, 10)] <- NA
+    iris$Sepal.Width[c(26, 43, 56, 11, 66, 132, 23, 133, 131, 28)] <- NA
[…]
