+++ tests/testthat/test-printing.R
@@ -11,23 +11,23 @@ if (require("testthat") && require("effectsize")) {
 
 
   test_that("effectsize difference", {
-    d <- cohens_d(1:3, c(1,1:3))
+    d <- cohens_d(1:3, c(1, 1:3))
     testthat::expect_output(print(d), regexp = "Cohen")
     testthat::expect_output(print(d), regexp = "pooled")
     testthat::expect_output(print(d, append_CL = TRUE), regexp = "U3")
 
-    d <- cohens_d(1:3, c(1,1:3), pooled_sd = FALSE)
+    d <- cohens_d(1:3, c(1, 1:3), pooled_sd = FALSE)
     testthat::expect_output(print(d), regexp = "un-pooled")
 
-    d <- hedges_g(1:3, c(1,1:3), correction = 1)
+    d <- hedges_g(1:3, c(1, 1:3), correction = 1)
     testthat::expect_output(print(d), regexp = "Hedges and Olkin")
 
-    d <- hedges_g(1:3, c(1,1:3), correction = 2)
+    d <- hedges_g(1:3, c(1, 1:3), correction = 2)
     testthat::expect_output(print(d), regexp = "Hunter and Schmidt")
   })
 
   test_that("equivalence test effectsize", {
-    d <- cohens_d(1:3, c(1,1:3))
+    d <- cohens_d(1:3, c(1, 1:3))
     equtest <- equivalence_test(d)
     testthat::expect_output(print(equtest), regexp = "ROPE")
   })
@@ -42,6 +42,4 @@ if (require("testthat") && require("effectsize")) {
     r2 <- rules(1:3, letters[1:3])
     testthat::expect_output(print(r2), regexp = "values")
   })
-
-
 }
