### 2020-12-07 #2815 @jkarreth
Just a quick heads up that I will start my review in the next couple of days. I'm therefore un-checking the boxes in my section of the checklist. If I find anything that needs to be addressed, I will create a new issue in the [effectsize repository](https://github.com/easystats/effectsize) and link to those issues here.
