+++ NEWS.md
@@ -1,3 +1,5 @@
+# effectsize 0.4.1.1
+
 # effectsize 0.4.1
 
 ## Breaking Changes
