### 2020-12-07 567d6b1
reduce example runtime
