+++ R/standardize_parameters.R
@@ -114,13 +114,14 @@
 #' }
 #'
 #'
-#'
+#' \dontrun{
 #' if (require("rstanarm")) {
 #'   model <- stan_glm(Sepal.Length ~ Species + Petal.Width, data = iris, refresh = 0)
 #'   # standardize_posteriors(model, method = "refit")
 #'   # standardize_posteriors(model, method = "posthoc")
 #'   # standardize_posteriors(model, method = "smart")
 #'   head(standardize_posteriors(model, method = "basic"))
+#' }
 #' }
 #' }
 #'
