+++ man/eta_squared.Rd
@@ -222,6 +222,7 @@ if (require(lmerTest, quietly = TRUE)) {
 
 ## Bayesian Models (PPD)
 ## ---------------------
+\dontrun{
 if (require(rstanarm) && require(bayestestR) && require(car)) {
   fit_bayes <- stan_glm(mpg ~ factor(cyl) * wt + qsec,
     data = mtcars,
@@ -241,6 +242,7 @@ if (require(rstanarm) && require(bayestestR) && require(car)) {
   )
   aov_table <- car::Anova(fit_freq, type = 3)
   eta_squared(aov_table)
+}
 }
 }
 
