+++ man/standardize_parameters.Rd
@@ -164,13 +164,14 @@ if (require("lme4")) {
 }
 
 
-
+\dontrun{
 if (require("rstanarm")) {
   model <- stan_glm(Sepal.Length ~ Species + Petal.Width, data = iris, refresh = 0)
   # standardize_posteriors(model, method = "refit")
   # standardize_posteriors(model, method = "posthoc")
   # standardize_posteriors(model, method = "smart")
   head(standardize_posteriors(model, method = "basic"))
+}
 }
 }
 
