### 2020-12-08 b4cee6c
More verbose `verbose` arg
#225
