+++ R/eta_squared.R
@@ -17,9 +17,10 @@
 #'   variables are manipulated. Can also be a character vector of observed
 #'   (non-manipulated) variables, in which case generalized Eta Squared is
 #'   calculated taking these observed variables into account. For `afex_aov`
-#'   model, `generalized = TRUE`, the observed variables are extracted
+#'   model, when `generalized = TRUE`, the observed variables are extracted
 #'   automatically from the fitted model, if they were provided then.
 #' @inheritParams chisq_to_phi
+#' @inheritParams standardize
 #' @param ... For Bayesian models, passed to `ss_function`. Otherwise ignored.
 #'
 #' @return
@@ -191,8 +192,9 @@ eta_squared <- function(model,
                         partial = TRUE,
                         generalized = FALSE,
                         ci = 0.9,
+                        verbose = TRUE,
                         ...) {
-  out <- .anova_es(model, type = "eta", partial = partial, generalized = generalized, ci = ci)
+  out <- .anova_es(model, type = "eta", partial = partial, generalized = generalized, ci = ci, verbose = verbose)
   class(out) <- unique(c("effectsize_table", "see_effectsize_table", class(out)))
   return(out)
 }
@@ -202,8 +204,9 @@ eta_squared <- function(model,
 omega_squared <- function(model,
                           partial = TRUE,
                           ci = 0.9,
+                          verbose = TRUE,
                           ...) {
-  out <- .anova_es(model, type = "omega", partial = partial, ci = ci)
+  out <- .anova_es(model, type = "omega", partial = partial, ci = ci, verbose = verbose)
   class(out) <- unique(c("effectsize_table", "see_effectsize_table", class(out)))
   return(out)
 }
@@ -213,8 +216,9 @@ omega_squared <- function(model,
 epsilon_squared <- function(model,
                             partial = TRUE,
                             ci = 0.9,
+                            verbose = TRUE,
                             ...) {
-  out <- .anova_es(model, type = "epsilon", partial = partial, ci = ci)
+  out <- .anova_es(model, type = "epsilon", partial = partial, ci = ci, verbose = verbose)
   class(out) <- unique(c("effectsize_table", "see_effectsize_table", class(out)))
   return(out)
 }
@@ -225,14 +229,16 @@ epsilon_squared <- function(model,
 #'   returns the effect size for R-squared-change between the two models.
 #' @export
 cohens_f <- function(model, partial = TRUE, ci = 0.9, squared = FALSE,
+                     verbose = TRUE,
                      model2 = NULL, ...) {
   if (!is.null(model2)) {
-    return(.cohens_f_delta(model, model2, ci = ci, squared = squared))
+    return(.cohens_f_delta(model, model2, ci = ci, squared = squared, verbose = verbose))
   }
 
   res <- eta_squared(model,
     partial = partial,
-    ci = ci
+    ci = ci,
+    verbose = verbose
   )
 
   if ("Eta2_partial" %in% colnames(res)) {
@@ -263,8 +269,9 @@ cohens_f <- function(model, partial = TRUE, ci = 0.9, squared = FALSE,
 #' @rdname eta_squared
 #' @export
 cohens_f_squared <- function(model, partial = TRUE, ci = 0.9, squared = TRUE,
+                             verbose = TRUE,
                              model2 = NULL, ...) {
-  cohens_f(model, partial = partial, ci = ci, squared = squared, model2 = model2)
+  cohens_f(model, partial = partial, ci = ci, squared = squared, verbose = verbose, model2 = model2)
 }
 
 
@@ -277,6 +284,7 @@ cohens_f_squared <- function(model, partial = TRUE, ci = 0.9, squared = TRUE,
            partial = TRUE,
[…]
