### 2020-12-08 57d6b88
Revert change to warning, and fix tests.
