+++ R/standardize_info.R
@@ -301,9 +301,11 @@ standardize_info <- function(model, robust = FALSE, two_sd = FALSE, include_pseu
 #' @importFrom parameters check_heterogeneity demean
 #' @importFrom stats as.formula sd
 .std_info_pseudo <- function(model, params, model_matrix, types, robust = FALSE, two_sd = FALSE) {
-  warning("'robust' standardization not available for 'pseudo' method.",
-          call. = FALSE
-  )
+  if (robust) {
+    warning("'robust' standardization not available for 'pseudo' method.",
+            call. = FALSE
+    )
+  }
 
   f <- if (two_sd) 2 else 1
 
