+++ man/adjust.Rd
@@ -36,6 +36,7 @@ This function can be used to adjust the data for the effect of other variables p
 \examples{
 adjust(iris, effect = "Species", select = "Sepal.Length")
 \donttest{
+\dontrun{
 adjust(iris, effect = "Species", select = "Sepal.Length", multilevel = TRUE)
 adjust(iris, effect = "Species", select = "Sepal.Length", bayesian = TRUE)
 adjust(iris, effect = "Petal.Width", select = "Sepal.Length", additive = TRUE)
@@ -48,6 +49,7 @@ adjust(iris,
   multilevel = TRUE, additive = TRUE
 )
 adjust(iris)
+}
 }
 
 }
