### 2020-12-09 easystats/effectsize#227 @jkarreth
JOSS review #2
@mattansb This is the second review of the JOSS submission. The first review by @trashbirdecology was already amazingly thorough, so much credit to @trashbirdecology for working through the vast majority of potential issues.

The review here is split in three parts: (1) issues I'd ask you to fix or explain, (2) issues I suggest considering before the article is accepted, and (3) thoughts for future iterations of the project.

I really like this package. I'm not an easystats user, but am very impressed by how well connected and well-thought-out the easystats ecosystem is. Fantastic effort!

# Issues I'd ask you to fix or explain

- [x] The function call to the cars example in the JOSS paper has an extra comma. Not an error, but consider correcting for aesthetic reasons.

- [x] The address & publisher for Cohen (1988) seem odd. Please correct to: New York: Academic Press, Inc. (via [here](https://books.google.com/books?hl=en&lr=&id=rEe0BQAAQBAJ&oi=fnd&pg=PP1&ots=swTXMuMXqc&sig=ZiRmMqmKOgEZnBppW-grZrWI4qA#v=onepage&q&f=false))

- [x] In [the `standardize_parameters` vignette](https://easystats.github.io/effectsize/articles/standardize_parameters.html), I'd recommend using `parameters::model_parameters`, like in the other vignettes.

- [x] This is probably a dumb question and reveals my ignorance, but why do most functions not return a standard error? E.g. `lm(Sepal.Length ~ Petal.Length, data = iris) %>%
		standardize_parameters`
could return a standard error in addition to CIs? It's OK if this is a choice by the package developers; I'm just thinking that many novice users would look for SEs, too.

- [x] Following the [other review](https://github.com/easystats/effectsize/issues/196), I thought the package main branch is now called "main" instead of "master", but "master" still exists and appears ahead of "main". Is that intended? I see that issue #201 is still open, so you're most likely onto that?

- [x] In the JOSS paper's first paragraph, you might consider adding a reference to latent variables around the "(especially when units of measurement are not meaningful)". Standardized effect sizes are particularly useful when using latent factor scores, for example. Adding this might clarify the package's appeal to some additional readers.

# Issues I suggest considering before the article is accepted

- [x] It would be nice if the [`convert` vignette](https://easystats.github.io/effectsize/articles/convert.html#converting-between-r-and-d) had the first example data available to insert/reproduce, i.e. create it within the vignette.

- [x] The package help files ould describe the value of data frames that results from function calls better. Maybe an example would help the reader quickly use this data frame for further processing. This could also help with something like combining functions, such as a hypothetical call to `interpret_d(standardize_parameters(mod)$Std_Median, rules = "cohen1988")`.

- [x] the paper could also mention the `scale()` function to standardize a matrix. That's what I was taught in my first R course many, many years ago :)

# Thoughts for future iterations of the project

- [x] Consider replacing examples using the `iris` data with another dataset, e.g. `mtcars` or [`penguins`](https://allisonhorst.github.io/palmerpenguins/). See [here](https://www.garrickadenbuie.com/blog/lets-move-on-from-iris/) for why.

- [x] A nice-to-have from my perspective would be if the output from **effectsize** could easily be passed on to other packages that allow automated output in the form of regression tables (texreg, stargazer, broom, etc.).
