### 2020-12-09 #2815 @jkarreth
@mattansb I created the review as an issue over at https://github.com/easystats/effectsize/issues/227. Thanks to @trashbirdecology 's thorough review, there wasn't much left for me to find. I added a few suggestions you might consider beyond the JOSS paper. Still trying to check the functionality of all parts of the package, but am close to done.
