+++ R/standardize_info.R
@@ -6,6 +6,11 @@
 #' @param include_pseudo (For (G)LMMs) Should Pseudo-standardized information be included?
 #' @param ... Arguments passed to or from other methods.
 #'
+#' @return A data frame with information on each parameter (see
+#'   [parameters::parameters_type]), and various standardization coefficients
+#'   for the post-hoc methods (see [standardize_parameters()]) for the predictor
+#'   and the response.
+#'
 #' @family standardize
 #'
 #' @examples
