+++ R/xtab.R
@@ -43,7 +43,9 @@
 #' @inheritSection cohens_d Confidence Intervals
 #' @inheritSection chisq_to_phi CI Contains Zero
 #'
-#' @return A data frame with the effect size(s), and confidence interval(s).
+#' @return A data frame with the effect size (`Cramers_v`, `phi` (possibly with
+#'   the suffix `_adjusted`), `Odds_ratio`, `Risk_ratio` (possibly with the
+#'   prefix `log_`), or `Cohens_g`) and its CIs (`CI_low` and `CI_high`).
 #'
 #' @seealso [chisq_to_phi()] for details regarding estimation and CIs.
 #' @family effect size indices
