+++ R/effectsize.R
@@ -24,6 +24,9 @@
 #' **For statistical models it is recommended to directly use the listed
 #' functions, for the full range of options they provide.**
 #'
+#' @return A data frame with the effect size (depending on input) and and its
+#'   CIs (`CI_low` and `CI_high`).
+#'
 #' @family effect size indices
 #'
 #' @examples
