### 2020-12-09 6075bae
Changes to vignettes
#227
