+++ vignettes/standardize_parameters.Rmd
@@ -62,11 +62,9 @@ standardize_parameters(m)
 Standardizing the coefficient of this *simple* linear regression gives a value of `0.87`, but did you know that for a simple regression this is actually the **same as a correlation**? Thus, you can eventually apply some (*in*)famous interpretation guidelines (e.g., Cohen's rules of thumb).
 
 ```{r}
-library(parameters)
-
 r <- cor.test(iris$Sepal.Length, iris$Petal.Length)
 
-model_parameters(r)
+parameters::model_parameters(r)
 ```
 
 
@@ -165,7 +163,7 @@ However, not all hope is lost yet - we can still try and recover the partial cor
 
 
 ```{r}
-params <- model_parameters(mod)
+params <- parameters::model_parameters(mod)
 
 t_to_r(params$t[-1], df_error = params$df_error[-1])
 ```
@@ -184,7 +182,7 @@ hardlyworking$age_g <- cut(hardlyworking$age,
 mod <- lm(salary ~ xtra_hours + n_comps + age_g + seniority,
           data = hardlyworking)
 
-model_parameters(mod)
+parameters::model_parameters(mod)
 ```
 
 It seems like the best or most important predictor is `n_comps` as it has the coefficient. However, it is hard to compare among predictors, as they are on different scales. To address this issue, we must have all the predictors on the same scale - usually in the arbitrary unit of *standard deviations*.
@@ -217,7 +215,7 @@ standardize_parameters(mod, method = "refit", two_sd = TRUE)
 mod_z <- standardize(mod, two_sd = FALSE, robust = FALSE)
 mod_z
 
-model_parameters(mod_z)
+parameters::model_parameters(mod_z)
 ```
 
 
@@ -258,8 +256,7 @@ Linear mixed models (LMM/HLM/MLM) offer an additional conundrum to standardizati
 The solution: standardize according to level of the predictor [@hoffman2015longitudinal, page 342]! Level 1 parameters are standardized according to variance *withing* groups, while level 2 parameters are standardized according to variance *between* groups. The resulting standardized coefficient are also called *pseudo*-standardized coefficients.[^Note that like method `"basic"`, these are based on the model matrix.]
 
 ```{r, eval=knitr::opts_chunk$get("eval") && require(lme4) && require(lmerTest)}
-library(lme4)
-m <- lmer(mpg ~ cyl + am + vs + (1|cyl), mtcars)
+m <- lme4::lmer(mpg ~ cyl + am + vs + (1|cyl), mtcars)
 
 standardize_parameters(m, method = "pseudo", df_method = "satterthwaite")
 
