+++ paper/paper.bib
@@ -127,7 +127,7 @@
   title={Statistical power analysis for the behavioral sciences, 2nd edn. {\'A}/L},
   author={Cohen, J},
   year={1988},
-  publisher={Erbaum Press, Hillsdale, NJ, USA}
+  publisher={New York: Academic Press, Inc.}
 }
 
 @article{hedges1985statistical,
