### 2020-12-09 d74fb6c
Update from_test_statistics.Rmd
