+++ R/cohens_d.R
@@ -47,8 +47,10 @@
 #' @family effect size indices
 #'
 #' @examples
-#' cohens_d(iris$Sepal.Length, iris$Sepal.Width)
-#' hedges_g("Sepal.Length", "Sepal.Width", data = iris)
+#' cohens_d(sleep$extra, sleep$group)
+#' hedges_g("extra", "group", data = sleep)
+#'
+#' cohens_d(sleep$extra[sleep$group==1], sleep$extra[sleep$group==2], paired = TRUE)
 #'
 #' cohens_d(mpg ~ am, data = mtcars)
 #' cohens_d(mpg ~ am, data = mtcars, pooled_sd = FALSE)
