+++ R/normalize.R
@@ -18,7 +18,7 @@
 #' normalize(c(0, 1, 5, -5, -2))
 #' normalize(c(0, 1, 5, -5, -2), include_bounds = FALSE)
 #'
-#' head(normalize(iris))
+#' head(normalize(trees))
 #' @references
 #' - Smithson M, Verkuilen J (2006). A Better Lemon Squeezer? Maximum-Likelihood Regression with Beta-Distributed Dependent Variables. Psychological Methods, 11(1), 54–71.
 #'
