+++ man/interpret_parameters.Rd
@@ -39,7 +39,3 @@ interpret_parameters(model, ...)
 \description{
 Automated interpretation of standardized slopes.
 }
-\examples{
-model <- lm(Sepal.Length ~ Species * Petal.Width, data = iris)
-interpret_parameters(model)
-}
