+++ vignettes/standardize_parameters.Rmd
@@ -54,7 +54,7 @@ Critically, **parameters standardization is not a trivial process**. Different t
 ```{r}
 library(effectsize)
 
-m <- lm(Sepal.Length ~ Petal.Length, data = iris)
+m <- lm(rating ~ complaints, data = attitude)
 
 standardize_parameters(m)
 ```
@@ -62,9 +62,9 @@ standardize_parameters(m)
 Standardizing the coefficient of this *simple* linear regression gives a value of `0.87`, but did you know that for a simple regression this is actually the **same as a correlation**? Thus, you can eventually apply some (*in*)famous interpretation guidelines (e.g., Cohen's rules of thumb).
 
 ```{r}
-r <- cor.test(iris$Sepal.Length, iris$Petal.Length)
+r <- cor.test(attitude$rating, attitude$complaints)
 
-parameters::model_parameters(r)
+effectsize(r)
 ```
 
 
