### 2020-12-09 easystats/effectsize#227 @mattansb
Hi @jkarreth !

Thanks for the kind words!

I have responded to your comments below. Let me know if I've missed anything or miss-addressed anything.
All changes are in the #228 PR.

Thanks!

# Issues I'd ask you to fix or explain

> - [ ] The function call to the cars example in the JOSS paper has an extra comma. Not an error, but consider correcting for aesthetic reasons.

Fixed!
https://github.com/easystats/effectsize/pull/228/commits/94744d8afff13d28aaf69c6aba0230585553c9ff#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424L112

> - [ ] The address & publisher for Cohen (1988) seem odd. Please correct to: New York: Academic Press, Inc. (via [here](https://books.google.com/books?hl=en&lr=&id=rEe0BQAAQBAJ&oi=fnd&pg=PP1&ots=swTXMuMXqc&sig=ZiRmMqmKOgEZnBppW-grZrWI4qA#v=onepage&q&f=false))

Done (https://github.com/easystats/effectsize/pull/228/commits/5d38a77e4825dc1e0752cf06c1235e46675be721).

> - [ ] In [the `standardize_parameters` vignette](https://easystats.github.io/effectsize/articles/standardize_parameters.html), I'd recommend using `parameters::model_parameters`, like in the other vignettes.

Oh, good suggestion - fixed!
https://github.com/easystats/effectsize/pull/228/commits/6075bae79b0f93ca9bb415a29c2f1c9216afdaf8#diff-40521620de914f93819a37dd1d1705cdb4fc8b341fcee916f2cbd02ca5a128ffR67

> - [ ] This is probably a dumb question and reveals my ignorance, but why do most functions not return a standard error? E.g. `lm(Sepal.Length ~ Petal.Length, data = iris) %>%
		standardize_parameters`
could return a standard error in addition to CIs? It's OK if this is a choice by the package developers; I'm just thinking that many novice users would look for SEs, too.

Good question - I've tried to keep the output of all function to the _Effect size + CI_ format, as not all effect sizes have accompanying SEs (at least that is not the method we've used for the CI estimation). However, specifically in `standardize_parameters()`, the SEs are actually retuned as an attribute:

``` r
library(magrittr)
library(effectsize)

(Z <- lm(Sepal.Length ~ Petal.Length, data = iris) %>%
  standardize_parameters())
#> Parameter    | Coefficient (std.) |        95% CI
#> -------------------------------------------------
#> (Intercept)  |          -5.03e-16 | [-0.08, 0.08]
#> Petal.Length |               0.87 | [ 0.79, 0.95]
#>
#> # Standardization method: refit

attr(Z, "standard_error")
#> [1] 0.04013870 0.04027317
```

I've made this explicit in the docs (under `return`). Thanks!
https://github.com/easystats/effectsize/pull/228/commits/3a3e29148e2aec7f190681478be1c62b3e07a352#diff-1d1ffd135ac6eac822153973d2d9ec0c81e6812576fe2170aa3c6d4263069ff0R84-R86

> - [ ] Following the [other review](https://github.com/easystats/effectsize/issues/196), I thought the package main branch is now called "main" instead of "master", but "master" still exists and appears ahead of "main". Is that intended? I see that issue #201 is still open, so you're most likely onto that?

Once the review is over (🤞) we will switch to main ( #210 ). I just didn't want to mess with the linking to JOSS.

> - [ ] In the JOSS paper's first paragraph, you might consider adding a reference to latent variables around the "(especially when units of measurement are not meaningful)". Standardized effect sizes are particularly useful when using latent factor scores, for example. Adding this might clarify the package's appeal to some additional readers.

Hmm, do you have any reference in mind? I've changed to "_(especially when units of measurement are not meaningful, e.g. in the use of estimated latent variables)_".
https://github.com/easystats/effectsize/pull/228/files?file-filters%5B%5D=.md#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R38

# Issues I suggest considering before the article is accepted

> - [ ] It would be nice if the [`convert` vignette](https://easystats.github.io/effectsize/articles/convert.html#converting-between-r-and-d) had the first example data available to insert/reproduce, i.e. create it within the vignette.

I don't even know why it was `echo=FALSE` here! Fixed!
https://github.com/easystats/effectsize/pull/228/commits/6075bae79b0f93ca9bb415a29c2f1c9216afdaf8#diff-ef85b163a761a7c29eda485a4d3583278bc70e7d7e97d7d3a7566e1ecfb0c97fL44

> - [ ] The package help files ould describe the value of data frames that results from function calls better. Maybe an example would help the reader quickly use this data frame for further processing. This could also help with something like combining functions, such as a hypothetical call to `interpret_d(standardize_parameters(mod)$Std_Median, rules = "cohen1988")`.

I've added more verbose information for the expected column names and such. Let me know if this is satisfactory. (https://github.com/easystats/effectsize/pull/228/commits/3a3e29148e2aec7f190681478be1c62b3e07a352)

> - [ ] the paper could also mention the `scale()` function to standardize a matrix. That's what I was taught in my first R course many, many years ago :)

Great suggestion!!! Added to `# Comparison to Other Packages`
https://github.com/easystats/effectsize/pull/228/commits/94744d8afff13d28aaf69c6aba0230585553c9ff#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R46

# Thoughts for future iterations of the project

> - [ ] Consider replacing examples using the `iris` data with another dataset, e.g. `mtcars` or [`penguins`](https://allisonhorst.github.io/palmerpenguins/). See [here](https://www.garrickadenbuie.com/blog/lets-move-on-from-iris/) for why.

This was gradually happening anyway, but now it's final - not a single `iris` in any of the docs / vignettes! (https://github.com/easystats/effectsize/pull/228/commits/b250fe7808ef5f54f0a49a298c251b11c08ee8d5)

> - [ ] A nice-to-have from my perspective would be if the output from **effectsize** could easily be passed on to other packages that allow automated output in the form of regression tables (texreg, stargazer, broom, etc.).

This is something that is in the works with `papaja` (see #156), and the `report` package (to hopefully be released soon)!
