+++ man/effectsize-CIs.Rd
@@ -22,10 +22,11 @@ For positive only effect sizes (Eta squared, Cramer's V, etc.; Effect sizes
 associated with Chi-squared and F distributions), special care should be
 taken when interpreting CIs with a lower bound equal to 0, and even more care
 should be taken when the \emph{upper} bound is equal to 0 (Steiger, 2004; Morey et
-al., 2016). For example:\preformatted{eta_squared(aov(mpg ~ factor(gear), mtcars[1:5, ]))
-#> Parameter    | Eta2 |       90\% CI
-#> ----------------------------------
-#> factor(gear) | 0.33 | [0.00, 0.74]
+al., 2016). For example:\if{html}{\out{<div class="r">}}\preformatted{eta_squared(aov(mpg ~ factor(gear) + factor(cyl), mtcars[1:7, ]))
+}\if{html}{\out{</div>}}\preformatted{## Parameter    | Eta2 (partial) |       90\% CI
+## --------------------------------------------
+## factor(gear) |           0.58 | [0.00, 0.84]
+## factor(cyl)  |           0.46 | [0.00, 0.78]
 }
 }
 
@@ -33,15 +34,14 @@ al., 2016). For example:\preformatted{eta_squared(aov(mpg ~ factor(gear), mtcars
 
 For very large sample sizes, the width of the CI can be smaller than the
 tolerance of the optimizer, resulting in CIs of width 0. This can also,
-result in the estimated CIs excluding the point estimate. For example:\preformatted{chisq_to_cramers_v(13223.73, n = 76227, nrow = 6, ncol = 1)
-#> Cramer's V |       95\% CI
-#> -------------------------
-#>       0.19 | [0.20, 0.20]
-
-t_to_d(80, df_error = 4555555)
-#>    d |       95\% CI
-#> -------------------
-#> 0.07 | [0.08, 0.08]
+result in the estimated CIs excluding the point estimate. For example:\if{html}{\out{<div class="r">}}\preformatted{chisq_to_cramers_v(13223.73, n = 76227, nrow = 6, ncol = 1)
+}\if{html}{\out{</div>}}\preformatted{## Cramer's V |       95\% CI
+## -------------------------
+##       0.19 | [0.20, 0.20]
+}\if{html}{\out{<div class="r">}}\preformatted{t_to_d(80, df_error = 4555555)
+}\if{html}{\out{</div>}}\preformatted{##    d |       95\% CI
+## -------------------
+## 0.07 | [0.08, 0.08]
 }
 }
 
