### 2020-12-10 easystats/effectsize#227 @jkarreth
@mattansb , Thanks for the quick turnaround! I'll check off issues in my review here in this issue to let you know what's addressed. Specific responses below:

>> In the JOSS paper's first paragraph, you might consider adding a reference to latent variables around the "(especially when units of measurement are not meaningful)". Standardized effect sizes are particularly useful when using latent factor scores, for example. Adding this might clarify the package's appeal to some additional readers.

> Hmm, do you have any reference in mind? I've changed to "(especially when units of measurement are not meaningful, e.g. in the use of estimated latent variables)".
https://github.com/easystats/effectsize/pull/228/files?file-filters%5B%5D=.md#diff-e504eb580b095a7e65428b098183a581e475f0fb316db95287eacd7d4f344424R38

Quick suggestion (others might have better ideas?): p. 165 in Bollen, K. A. (1989). Structural Equations with Latent Variables. New York: Wiley

>>This is probably a dumb question and reveals my ignorance, but why do most functions not return a standard error? E.g. lm(Sepal.Length ~ Petal.Length, data = iris) %>% standardize_parameters
could return a standard error in addition to CIs? It's OK if this is a choice by the package developers; I'm just thinking that many novice users would look for SEs, too.

> Good question - I've tried to keep the output of all function to the Effect size + CI format, as not all effect sizes have accompanying SEs (at least that is not the method we've used for the CI estimation). However, specifically in standardize_parameters(), the SEs are actually retuned as an attribute:
> ...

OK! Maybe you could put a note to this point in the JOSS paper so it's easier for readers to see? Just a suggestion - not a condition for me closing this review.

>> The package help files could describe the value of data frames that results from function calls better. Maybe an example would help the reader quickly use this data frame for further processing. This could also help with something like combining functions, such as a hypothetical call to `interpret_d(standardize_parameters(mod)$Std_Median, rules = "cohen1988")`.

> I've added more verbose information for the expected column names and such. Let me know if this is satisfactory. (3a3e291)

Yup, that's helpful! Resolved.
