### 2020-12-10 easystats/effectsize#227 @jkarreth
One open issue from the initial review:

-[ ] .bib entry for `@cohen1988statistical` still has `{\'A}/L}` in the title field - I think that can go?
