### 2020-12-10 f44a3fa
Fix Cohen (1988) citation
#227
