+++ paper/paper.bib
@@ -124,10 +124,10 @@
 }
 
 @book{cohen1988statistical,
-  title={Statistical power analysis for the behavioral sciences, 2nd edn. {\'A}/L},
+  title={Statistical power analysis for the behavioral sciences, 2nd ed.},
   author={Cohen, J},
   year={1988},
-  publisher={New York: Academic Press, Inc.}
+  publisher={New York: Routledge}
 }
 
 @article{hedges1985statistical,
