+++ R/convert_d_to_common_language.R
@@ -20,7 +20,7 @@
 #' @family convert between effect sizes
 #'
 #' @references
-#' - Cohen, J. (1977). Statistical power analysis for the behavioral sciencies. Routledge.
+#' - Cohen, J. (1977). Statistical power analysis for the behavioral sciences. New York: Routledge.
 #' - Reiser, B., & Faraggi, D. (1999). Confidence intervals for the overlapping coefficient: the normal equal variance case. Journal of the Royal Statistical Society, 48(3), 413-418.
 #' - Ruscio, J. (2008). A probability-based measure of effect size: robustness to base rates and other factors. Psychological methods, 13(1), 19–30.
 #'
