+++ man/cohens_d.Rd
@@ -105,7 +105,7 @@ print(cohens_d(mpg ~ am, data = mtcars), append_CL = TRUE)
 }
 \references{
 \itemize{
-\item Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.
+\item Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd Ed.). New York: Routledge.
 \item Hedges, L. V. & Olkin, I. (1985). Statistical methods for meta-analysis. Orlando, FL: Academic Press.
 \item Hunter, J. E., & Schmidt, F. L. (2004). Methods of meta-analysis: Correcting error and bias in research findings. Sage.
 \item McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
