+++ paper/paper.bib
@@ -143,3 +143,10 @@
   year={2014},
   publisher={Academic press}
 }
+
+@article{bollen1989structural,
+  title={Structural equations with latent variables Wiley},
+  author={Bollen, Kenneth A},
+  journal={New York},
+  year={1989}
+}
