### 2020-12-10 easystats/effectsize#227 @mattansb
- Fixed the Cohen (1988) citation (in the paper and across the whole package).
- Added bollen1989structural ref

> OK! Maybe you could put a note to this point in the JOSS paper so it's easier for readers to see? Just a suggestion - not a condition for me closing this review.

I think that is covered by the first line in the "Examples of Features": _**effectsize** provides various functions for extracting and estimating effect sizes and their confidence intervals_ (and is evident in the code examples)? Is there any particular place you think this should be emphasized?

---

When you close this issue, I'll merge the PR #228 , and have the new PDF generated on the JOSS issue.
Thanks for taking the time!
