### 2020-12-10 a6c12db
Minor printing fixes
