### 2020-12-10 4ad5b64
`glass_delta()` returns correct CIs
