+++ R/cohens_d.R
@@ -175,9 +175,8 @@ glass_delta <- function(x, y = NULL, data = NULL, ci = 0.95, correction) {
     n2 <- length(y)
     n <- n1 + n2
 
-    hn <- (1 / n1 + 1 / n2)
-    df <- n - 2
     if (type == "d" | type == "g") {
+      hn <- (1 / n1 + 1 / n2)
       if (pooled_sd) {
         s <- suppressWarnings(sd_pooled(x, y))
 
@@ -194,9 +193,11 @@ glass_delta <- function(x, y = NULL, data = NULL, ci = 0.95, correction) {
         df <- se^4 / (se1^4 / (n1 - 1) + se2^4 / (n2 - 1))
       }
     } else if (type == "delta") {
+      hn <- 1 / (n2 - 1)
       s <- stats::sd(y)
 
-      se <- s * sqrt(1 / n1 + 1 / n2)
+      se <- s / sqrt(n2)
+      df <- n2 - 1
     }
   }
 
