### 2020-12-10 easystats/effectsize#227 @jkarreth
> I think that is covered by the first line in the "Examples of Features": effectsize provides various functions for extracting and estimating effect sizes and their confidence intervals (and is evident in the code examples)? Is there any particular place you think this should be emphasized?

Maybe simply change to something like: *effectsize provides various functions for extracting and estimating effect sizes and measures of uncertainty (such as confidence intervals and standard errors)* ?

But I leave it to you; this is not a critical issue.

I'm closing the review here because all items above are addressed. Thanks! The only remaining item in my [review over at JOSS](https://github.com/openjournals/joss-reviews/issues/2815) is the "Functionality" checkpoint; I'll try a few more things in the next few days and will leave a note when that is checked off as well.
