### 2020-12-10 easystats/effectsize#196 @trashbirdecology
@mattansb Cohen's still isn't cited until the last reference to Cohen's d.
