+++ paper/paper.bib
@@ -150,3 +150,33 @@
   journal={New York},
   year={1989}
 }
+
+@book{cramer1946mathematical,
+  title={Mathematical Methods of Statistics},
+  author={Cram{\'e}r, Harald},
+  year={1946},
+  publisher={Princeton: Princeton University Press}
+}
+
+@article{kelley1935unbiased,
+  title={An unbiased correlation ratio measure},
+  author={Kelley, Truman L},
+  journal={Proceedings of the National Academy of Sciences of the United States of America},
+  volume={21},
+  number={9},
+  pages={554},
+  year={1935},
+  publisher={National Academy of Sciences}
+}
+
+
+@article{friedman1982simplified,
+  title={Simplified determinations of statistical power, magnitude of effect and research sample sizes},
+  author={Friedman, Herbert},
+  journal={Educational and Psychological Measurement},
+  volume={42},
+  number={2},
+  pages={521--526},
+  year={1982},
+  publisher={Sage Publications Sage CA: Thousand Oaks, CA}
+}
