### 2020-12-10 easystats/effectsize#196 @mattansb
@trashbirdecology How did I miss that??
Added (and a bunch more on the way!) - https://github.com/easystats/effectsize/pull/228/commits/0065daae7513afd7782b1c88d13c4ed7ea2cea9a
