### 2020-12-11 easystats/effectsize#227 @jkarreth
Sorry to reopen, but looking at the re-compiled paper I noticed that a few references need to be fixed in the .bib file.

- Some books are entered as `@article` instead of `@book`
- Capitalization for titles and journal name is off in a few entries (e.g. `Bmj`)
- Check `hedges1985statistical` and `bollen1989structural ` for a few issues
- If possible, please provide DOIs and ISBNs consistently.

Thanks for looking into that - and apologies for overlooking it earlier. Will close the issue here when I see a corrected .bib file.
