### 2020-12-11 e2ec3b9
Fix citations
#227
