+++ paper/paper.bib
@@ -1,4 +1,4 @@
-@Manual{rcore,
+@manual{rcore,
   title = {{R}: A Language and Environment for Statistical Computing},
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
@@ -8,12 +8,12 @@
 }
 
 @article{steiger2004beyond,
-  title={Beyond the F test: effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis.},
+  title={Beyond the F test: effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis},
   author={Steiger, James H},
-  journal={Psychological methods},
+  journal={Psychological Methods},
   volume={9},
   number={2},
-  pages={164},
+  pages={164--182},
   year={2004},
   publisher={American Psychological Association},
   doi = {10.1037/1082-989X.9.2.164}
@@ -24,23 +24,26 @@
   author={Hoffman, Lesa},
   year={2015},
   publisher={Routledge},
+  address={New York},
   ISBN = {9780415876025}
 }
 
-@article{borenstein2009converting,
+@inbook{borenstein2009converting,
   title={Converting among effect sizes},
-  author={Borenstein, Michael and Hedges, Larry V and Higgins, JPT and Rothstein, Hannah R},
-  journal={Introduction to meta-analysis},
-  pages={45--49},
+  booktitle={Introduction to meta-analysis},
+  author={Borenstein, Michael and Hedges, Larry V and Higgins, Julian PT and Rothstein, Hannah R},
   year={2009},
-  publisher={John Wiley \& Sons, Chichester},
+  chapter={7},
+  pages={45--49},
+  publisher={John Wiley \& Sons},
+  address={Chichester},
   doi = {10.1002/9780470743386.ch7}
 }
 
 @article{grant2014converting,
   title={Converting an odds ratio to a range of plausible relative risks for better communication of research findings},
   author={Grant, Robert L},
-  journal={Bmj},
+  journal={British Medical Journal},
   volume={348},
   pages={f7450},
   year={2014},
@@ -49,19 +52,19 @@
 }
 
 @article{olejnik2003generalized,
-  title={Generalized eta and omega squared statistics: measures of effect size for some common research designs.},
+  title={Generalized eta and omega squared statistics: measures of effect size for some common research designs},
   author={Olejnik, Stephen and Algina, James},
-  journal={Psychological methods},
+  journal={Psychological Methods},
   volume={8},
   number={4},
-  pages={434},
+  pages={434--447},
   year={2003},
   publisher={American Psychological Association},
   doi = {10.1037/1082-989X.8.4.434}
 }
 
 @article{ludecke2020extracting,
-  title={Extracting, Computing and Exploring the Parameters of Statistical Models using R},
+  title={Extracting, computing and exploring the parameters of statistical models using R},
   author={L{\"u}decke, Daniel and Ben-Shachar, Mattan S and Patil, Indrajeet and Makowski, Dominique},
   journal={Journal of Open Source Software},
   volume={5},
@@ -80,7 +83,7 @@
   doi = {10.5281/zenodo.2074621},
 }
 
-@Manual{sjoberg2020gtsummary,
+@manual{sjoberg2020gtsummary,
   title = {gtsummary: Presentation-Ready Data Summary and Analytic Result Tables},
   author = {Daniel D. Sjoberg and Michael Curry and Margie Hannum and Karissa Whiting and Emily C. Zabor},
   year = {2020},
@@ -89,7 +92,7 @@
 }
 
 @article{luedecke2019insight,
[…]
