### 2020-12-11 easystats/effectsize#227 @mattansb
@jkarreth I'm 99.9% I got all of them (see latest commit https://github.com/easystats/effectsize/commit/e2ec3b9d57cb1d2b8757abe5987b4e6caac2e7b8) - let me know if I somehow missed anything 😅
