### 2020-12-11 e5b6f2e
use insight for BF-prop-test
https://github.com/easystats/parameters/issues/362
