### 2020-12-11 fb788f9
Use insight for BF-xtab-test
