### 2020-12-11 easystats/effectsize#227 @jkarreth
yup, I think all looks good now! Closing the review again.
