### 2020-12-11 easystats/effectsize#227 @mattansb
Awesome - thanks again!
