### 2020-12-11 62f53b1
add `mu` arg to stad_diff
@IndrajeetPatil FYI
