### 2020-12-15 948d276
Update DESCRIPTION
https://github.com/easystats/easystats/issues/79#issuecomment-744632884
