### 2020-12-17 71a5256
Fix printing of paried cohen's d (not un/pooled sd)
