### 2020-12-18 7200c8c
add f_to_omega2 conversion
