### 2020-12-18 757a9a1
Clean up eta2_to_f2
