+++ NAMESPACE
@@ -98,11 +98,12 @@ export(d_to_r)
 export(effectsize)
 export(epsilon_squared)
 export(equivalence_test)
-export(eta2_to_F)
+export(eta2_to_f)
+export(eta2_to_f2)
 export(eta_squared)
 export(eta_squared_posterior)
+export(f2_to_eta2)
 export(f_to_eta2)
-export(f_to_omega2)
 export(format_standardize)
 export(glass_delta)
 export(hedges_g)
