+++ man/F_to_eta2.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/convert_tF_to_pve.R
+% Please edit documentation in R/convert_tF_to_anova.R
 \name{F_to_eta2}
 \alias{F_to_eta2}
 \alias{t_to_eta2}
