+++ R/convert_between_anova.R
@@ -17,6 +17,9 @@
 #' (such as Epsilon- or Omega-Squared), the resulting Cohen's *f* is likewise a
 #' less biaed estimate of Cohen's *f*.
 #'
+#' @seealso [eta_squared()] for more details.
+#' @family convert between effect sizes
+#'
 #' @references
 #' - Cohen, J. (1988). Statistical power analysis for the behavioral sciences (2nd Ed.). New York: Routledge.
 #' - Steiger, J. H. (2004). Beyond the F test: Effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis. Psychological Methods, 9, 164-182.
