+++ man/oddsratio_to_riskratio.Rd
@@ -39,6 +39,7 @@ Grant, R. L. (2014). Converting an odds ratio to a range of plausible relative r
 Other convert between effect sizes: 
 \code{\link{d_to_common_language}()},
 \code{\link{d_to_r}()},
+\code{\link{eta2_to_f2}()},
 \code{\link{odds_to_probs}()}
 }
 \concept{convert between effect sizes}
