### 2020-12-19 72a77e2
enhance and fix bugs in data arguments passed to cohens_d
