+++ NEWS.md
@@ -6,10 +6,12 @@
 
 ## Bug fixes
 
+- `cohens_d()` family / `sd_pooled()` now properly fails when given a missing column name.
 - `glass_delta()` returns correct CIs (was too narrow).
 
 ## Changes
 
+- `cohens_d()` family / `sd_pooled()` now respect any transformations (e.g. `I(log(x) - 3) ~ factor(y)`) in a passed formula.
 - `eta_squared()` family of functions gains a `verbose` argument.
 - `verbose` argument more strictly respected.
 
