+++ paper/paper.bib
@@ -141,7 +141,6 @@
   publisher={Academic Press},
   address={San Diego},
   ISBN={9780080570655},
-  doi={C2009-0-03396-0},
   year={1985}
 }
 
