### 2020-12-21 #2815 @mattansb
The first DOI is not a DOI, but an ISBN - not sure why whedon recognizes is at a DOI...?

The second DOI was removed - the book has an ISBN instead.

@mikldk let me know if this is sufficient.
