### 2020-12-21 #2815 @mikldk
@mattansb, can you try also entering `DOI = {10.1002/9781118619179}` for the book instead of just the ISBN? That should be the correct one according to <https://onlinelibrary.wiley.com/doi/book/10.1002/9781118619179>. (We like DOI's :smiley:.)
