### 2020-12-22 #2815 @mikldk
@mattansb Can you please rename the Github release to e.g. "0.4.1.1 (JOSS Paper)"?
