### 2020-12-22 #2815 @mikldk
@mattansb
Please **carefully** review the titles, please...:

Zenodo:
effectsize: Estimation and Interpretation of Effect Size Indices and Standardized Parameters

Paper:
effectsize: Estimation of Effect Size Indices and Standardized Parameters

Please correct this, and **carefully**, **carefully** chech titles, authors etc. And make a new release (with name containing version, e.g. "0.4.1.2 (JOSS Paper)") and post the version number and DOI here.
