### 2020-12-22 #2815 @mattansb
Okay, I made a new release, and carefully checked the title and authors on Zenodo to match with the paper.

Sorry for the mess, hope this is now finalized:

DOI [10.5281/zenodo.4384509](https://doi.org/10.5281/zenodo.4384509).
