### 2020-12-22 #2815 @mikldk
@arfon The paper changed its title since review issue was created. I have just edited it here at Github. Should I do something else, too, for the change to happen anywhere else?
