### 2020-12-22 48d8522
fix interpret_bf when include_value is TRUE
