### 2020-12-22 2d1f102
effectsize.htest uses actual data + added McNemar test
#224
