### 2020-12-22 53e631e
add verbose to `effectsize()`
https://github.com/easystats/parameters/issues/371
