### 2020-12-22 9782fbb
fix tests for retriving data in effectsize()
