### 2020-12-23 78e72dd
fix: maintain class for interpret_bf
Also changed the interpretation of BF=1 to "no evidence against or in favour of"
