### 2020-12-23 5617f60
minor case changes in bib
