+++ paper/paper.bib
@@ -8,7 +8,7 @@
 }
 
 @article{steiger2004beyond,
-  title={Beyond the F test: effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis},
+  title={Beyond the {F} test: effect size confidence intervals and tests of close fit in the analysis of variance and contrast analysis},
   author={Steiger, James H},
   journal={Psychological Methods},
   volume={9},
@@ -64,7 +64,7 @@
 }
 
 @article{ludecke2020extracting,
-  title={Extracting, computing and exploring the parameters of statistical models using R},
+  title={Extracting, computing and exploring the parameters of statistical models using {R}},
   author={L{\"u}decke, Daniel and Ben-Shachar, Mattan S and Patil, Indrajeet and Makowski, Dominique},
   journal={Journal of Open Source Software},
   volume={5},
@@ -84,7 +84,7 @@
 }
 
 @manual{sjoberg2020gtsummary,
-  title = {gtsummary: Presentation-Ready Data Summary and Analytic Result Tables},
+  title = {{g}tsummary: Presentation-Ready Data Summary and Analytic Result Tables},
   author = {Daniel D. Sjoberg and Michael Curry and Margie Hannum and Karissa Whiting and Emily C. Zabor},
   year = {2020},
   note = {R package version 1.3.5},
@@ -92,7 +92,7 @@
 }
 
 @article{luedecke2019insight,
-  title = {insight: A unified interface to access information from model objects in R},
+  title = {{i}nsight: A unified interface to access information from model objects in {R}},
   volume = {4},
   doi = {10.21105/joss.01412},
   number = {38},
@@ -103,7 +103,7 @@
 }
 
 @manual{behrendt2014lmbeta,
-  title = {lm.beta: Add Standardized Regression Coefficients to lm-Objects},
+  title = {{l}m.beta: Add Standardized Regression Coefficients to lm-Objects},
   author = {Stefan Behrendt},
   year = {2014},
   note = {R package version 1.5-1},
@@ -119,7 +119,7 @@
 }
 
 @manual{kelley2020MBESS,
-  title = {MBESS: The MBESS R Package},
+  title = {MBESS: The MBESS {R} Package},
   author = {Ken Kelley},
   year = {2020},
   note = {R package version 4.8.0},
