### 2020-12-23 #2815 @danielskatz
👋 @mattansb - I've suggested some small changes to the paper and bib in https://github.com/easystats/effectsize/pull/246 - please merge this if they are ok with you.  If not, let me know what you disagree with.
